### 2024-11-26 #674 @RayStick
mapmetadata
Submitting Author Name: Rachael Stickland
Submitting Author Github Handle: @RayStick
Other Package Authors Github handles: (comma separated, delete if none) @BatoolMM, @Rainiefantasy
Repository:  https://github.com/aim-rsf/mapmetadata
Version submitted:
Submission type: Standard

Version accepted: TBD

---

```
Package: mapmetadata
Title: Map health metadata onto predefined research domains
Version: 3.0.0
    person("Rachael", "Stickland", , "rstickland@turing.ac.uk", role = c("aut", "cre"),
    person("Batool", "Almarzouq", role = "ctb",
    person("Mahwish", "Mohammad", role = "ctb",
    person("Daniel", "Delbarre", role = "ctb",
    person("Nida", "Ziauddeen", role = "ctb",
  )
Description: Prior to gaining full access to health datasets, explore
    publicly available metadata and map metadata onto predefined research
    domains.  This package uses structural metadata files downloaded from
    the Health Data Research Gateway (https://healthdatagateway.org/en).
    In theory, any metadata file with the same structure as the files
    downloaded from this gateway can be used with this package, but the
    package has been developed and tested on metadata files from this
    gateway only.
    cli,
    gridExtra,
    htmlwidgets,
    plotly,
    tidyr
    devtools,
    mockery,
    withr
    knitr
```

_This package is related to data access, as interacting with health metadata can help a researcher/research group decide what datasets to access for their research, and be more informed when writing their data access request. It involves data validation checks as it checks for completeness of metadata, and visualizes this._

_Any users of health metadata, specifically for research projects that are using large population datasets, with many latent variables (research domains/concepts) and they need to investigate which variables in the datasets map onto their research domains of interest._

_Not that I am aware of_

_Not Applicable_

_Not Applicable_
