+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   browseMetadata: browseMetadata.html
   HealthMetadata: HealthMetadata.html
-last_built: 2024-11-26T17:08Z
+last_built: 2024-11-26T17:18Z
 urls:
   reference: https://aim-rsf.github.io/browseMetadata/reference
   article: https://aim-rsf.github.io/browseMetadata/articles
