### 2024-11-27 #674 @mpadge
@RayStick No worries. When you're ready, please just open this issue again (and not a new issue), call `@ropensci-review-bot check package` and things will move on from there.
