### 2024-11-27 ae5d396
Add example in concensus_on_mismatch
