### 2024-11-27 465f78e
indicate internal functions
