### 2024-11-27 22ea46f
add internal flags
