+++ man/valid_comparison.Rd
@@ -19,8 +19,9 @@ valid_comparison(input_1, input_2, severity, severity_text)
 Returns nothing if inputs are equal. If inputs are not equal, returns variable text depending on level of severity.
 }
 \description{
-This function is called within the map_metadata_compare function. \cr \cr
+Internal Function: This function is called within the map_metadata_compare function. \cr \cr
 It reads two inputs to see if they are equal. \cr \cr
 If the test is 'warning' status and inputs are not equal it gives warning but continues. \cr \cr
 If the test is 'danger' status and inputs are not equal it stops and exits, with error message. \cr \cr
 }
+\keyword{internal}
