+++ R/data_manipulation.R
@@ -64,13 +64,15 @@ count_empty_desc <- function(table_df, table_colname) {
 
 #' join_outputs
 #'
-#' This function is called within the map_metadata_compare function. \cr \cr
+#' Internal Function: This function is called within the map_metadata_compare function. \cr \cr
 #' Joins output dataframes from two sessions, on the column DataElement.
 #'
 #' @param session_1 Dataframe from session 1
 #' @param session_2 Dataframe from session 2
 #' @return Dataframe with information from session 1 and 2, joined on column DataElement.
 #' @importFrom dplyr left_join join_by select contains
+#' @keywords internal
+
 
 join_outputs <- function(session_1, session_2) {
   ses_join <- left_join(session_1, session_2, suffix = c("_ses1", "_ses2"), join_by(data_element))
