### 2024-11-27 1826eda
add examples to functions
