### 2024-11-27 2db7962
remove the examples to re-create them again after checking cmd check
