### 2024-11-27 acef2af
Update R/browse_metadata.R
Co-authored-by: Rachael Stickland <50215726+RayStick@users.noreply.github.com>
