### 2024-11-28 9fdeb83
add example for map_metadata_convert
