### 2024-11-28 28fa092
add example to map_metadata_compare
