+++ R/map_metadata_compare.R
@@ -5,7 +5,7 @@
 #'
 #' @param session_dir This directory should contain 2 csv files for each session (LOG_ and OUTPUT_), 4 csv files in total.
 #' @param session1_base Base file name for session 1 e.g. 'NationalCommunityChildHealthDatabase(NCCHD)_BLOOD_TEST_2024-07-05-16-07-38'
-#' @param session2_base Base file name for session 1 e.g. 'NationalCommunityChildHealthDatabase(NCCHD)_BLOOD_TEST_2024-07-08-12-03-30'
+#' @param session2_base Base file name for session 2 e.g. 'NationalCommunityChildHealthDatabase(NCCHD)_BLOOD_TEST_2024-07-08-12-03-30'
 #' @param json_file The full path to the metadata file used when running map_metadata (should be the same for session 1 and session 2)
 #' @param domain_file The full path to the domain file used when running map_metadata (should be the same for session 1 and session 2)
 #' @return It returns a csv output, which represents the consensus decisions between session 1 and session 2
@@ -13,6 +13,22 @@
 #' @importFrom utils read.csv write.csv
 #' @importFrom jsonlite fromJSON
 #' @importFrom cli cli_alert_success
+#' @examples
+#' # Locate file paths for the example files in the package
+#' session_dir <- system.file("outputs", package = "browseMetadata")
+#' session1_base <- "NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-11-27-14-19-55"
+#' session2_base <- "NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-11-27-14-23-52"
+#' json_file <- system.file("inputs", "national_community_child_health_database_(ncchd)_20240405T130125.json", package = "browseMetadata")
+#' domain_file <- system.file("inputs", "domain_list_demo.csv", package = "browseMetadata")
+#'
+#' # Run the function
+#' map_metadata_compare(
+#'   session_dir = session_dir,
+#'   session1_base = session1_base,
+#'   session2_base = session2_base,
+#'   json_file = json_file,
+#'   domain_file = domain_file
+#' )
 
 map_metadata_compare <- function(session_dir, session1_base, session2_base, json_file, domain_file) {
   timestamp_now_fname <- format(Sys.time(), "%Y-%m-%d-%H-%M-%S")
