+++ man/map_metadata_compare.Rd
@@ -31,6 +31,7 @@ This function is to be used after running the map_metadata function. \cr \cr
 It compares csv outputs from two sessions, finds their differences, and asks for a consensus. \cr \cr
 }
 \examples{
+\dontrun{
 # Locate file paths for the example files in the package
 session_dir <- system.file("outputs", package = "browseMetadata")
 session1_base <- "NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-11-27-14-19-55"
@@ -38,12 +39,12 @@ session2_base <- "NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-11-27-1
 json_file <- system.file("inputs", "national_community_child_health_database_(ncchd)_20240405T130125.json", package = "browseMetadata")
 domain_file <- system.file("inputs", "domain_list_demo.csv", package = "browseMetadata")
 
-# Run the function
+# Run the function - requires user interaction
 map_metadata_compare(
   session_dir = session_dir,
   session1_base = session1_base,
   session2_base = session2_base,
   json_file = json_file,
   domain_file = domain_file
-)
+)}
 }
