### 2024-11-28 81d99d3
simplify examples
