+++ R/map_metadata_convert.R
@@ -9,50 +9,12 @@
 #' @export
 #' @importFrom utils read.csv write.csv
 #' @examples
-#' # Load the browseMetadata package
-#' library(browseMetadata)
-#' # Define the get_output_paths function temporarily (for example)
-#' get_output_paths <- function(filename = NULL) {
-#'   # Resolve the output directory relative to the installed package
-#'   output_dir <- system.file("outputs", package = "browseMetadata")
+#' # Locate file path and file name for the example files in the package
+#' demo_output_dir <- system.file("outputs", package = "browseMetadata")
+#' demo_output_csv <- "OUTPUT_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-11-27-14-19-55.csv"
 #'
-#'   # Check if the output_dir was correctly found
-#'   if (output_dir == "") {
-#'     stop("The outputs directory could not be found. Make sure the package is installed correctly.")
-#'   }
-#'
-#'   # If a filename is provided, create the full file path (just the filename)
-#'   output_csv <- if (!is.null(filename)) {
-#'     filename  # Return just the filename (not full path yet)
-#'   } else {
-#'     NULL
-#'   }
-#'
-#'   # Return both output_dir and output_csv (just the filename, not full path)
-#'   list(output_dir = output_dir, output_csv = output_csv)
-#' }
-#'
-#' # Define input parameters using dynamic paths
-#' paths <- get_output_paths("OUTPUT_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-11-27-14-19-55.csv")
-#' output_csv <- paths$output_csv  # Just the filename
-#' output_dir <- paths$output_dir  # Directory path
-#'
-#' # Print paths for debugging (optional)
-#' print(output_csv)  # Check the filename
-#' print(output_dir)  # Check the directory path
-#'
-#' # Check if the directory exists before proceeding
-#' if (!dir.exists(output_dir)) {
-#'   stop("Output directory does not exist!")
-#' }
-#'
-#' # Run the function to convert the metadata
-#' map_metadata_convert(
-#'   output_csv = output_csv,  # Pass just the filename here
-#'   output_dir = output_dir   # Pass the directory path separately
-#' )
-
-
+#' # Run the function
+#' map_metadata_convert(output_csv = demo_output_csv, output_dir = demo_output_dir)
 map_metadata_convert <- function(output_csv, output_dir) {
   output <- read.csv(paste0(output_dir, "/", output_csv))
   output_long <- output[0, ] # make duplicate
