+++ _pkgdown.yml
@@ -23,11 +23,11 @@ reference:
 - title: Functions
 - subtitle: User functions
   contents:
-  - starts_with("map_metadata")
-  - browse_metadata
+    - starts_with("map_metadata")
+    - browse_metadata
 - subtitle: Internal functions
   contents:
- - @internal
+    - @internal
 
 
 
