+++ _pkgdown.yml
@@ -21,13 +21,10 @@ navbar:
     right: [search, github, lightswitch]
 reference:
 - title: Functions
-- subtitle: User functions
-  contents:
-    - starts_with("map_metadata")
-    - browse_metadata
-- subtitle: Internal functions
-  contents:
-    - @internal
-
-
-
+  - subtitle: User functions
+    contents:
+      - starts_with("map_metadata")
+      - browse_metadata
+  - subtitle: Internal functions
+    contents:
+      - "@internal"
