+++ _pkgdown.yml
@@ -20,12 +20,10 @@ navbar:
     left:  [intro, reference, articles, tutorials, news]
     right: [search, github, lightswitch]
 reference:
-  - title: Functions
+- title: "User functions"
+  contents:
+  - starts_with("map_metadata")
+  - browse_metadata
+  - title: "Internal Functions"
     contents:
-      - subtitle: User functions
-        contents:
-          - starts_with("map_metadata")
-          - browse_metadata
-      - subtitle: Internal functions
-        contents:
-          - "@internal"
+  - "@internal"
