+++ _pkgdown.yml
@@ -26,4 +26,4 @@ reference:
   - browse_metadata
 - title: "Internal Functions"
   contents:
-  - keywords: internal
+  - keywords: "internal"
