### 2024-12-06 f69b04d
has_keyword("internal")
