+++ _pkgdown.yml
@@ -22,8 +22,8 @@ navbar:
 reference:
 - title: "User functions"
   contents:
-    - starts_with("map_metadata")
-    - browse_metadata
+  - starts_with("map_metadata")
+  - browse_metadata
 - title: "Internal Functions"
   contents:
-    - keywords: "internal"
+  - has_keyword("internal")
