### 2024-12-06 39a2a88
unit test for map_metadata_convert
