### 2024-12-11 e16d329
remove old notes
