### 2024-12-11 7e0489a
unit test for map_metadata
