+++ R/map_metadata.R
@@ -45,6 +45,11 @@
 #' @importFrom utils packageVersion write.csv
 #' @importFrom ggplot2 ggsave
 
+# Wrap base function so that mock functions within unit tests are allowed
+readline_wrapper <- function(prompt) {
+  readline(prompt)
+  }
+
 map_metadata <- function(
     json_file = NULL,
     domain_file = NULL,
@@ -100,8 +105,8 @@ map_metadata <- function(
   ))
   cat("\n\n")
   cli_alert_info("Reference outputs from browse_metadata for information about the dataset")
-  cat("\nPress any key to continue ")
-  readline()
+  cat("\n")
+  readline_wrapper("Press any key to continue ")
 
   # WHICH TABLES FROM THE DATASET? ----
   ## Use 'user_prompt_list.R' to ask user which tables to process
@@ -146,7 +151,7 @@ map_metadata <- function(
       df_prev_exist <- FALSE
     }
 
-    table_note <- readline(paste(
+    table_note <- readline_wrapper(paste(
       "Optional free text note about this table",
       "(or press enter to continue): "
     ))
@@ -194,6 +199,7 @@ map_metadata <- function(
 
     output_df$timestamp <- timestamp_now
     output_df$table <- table_name
+
 
     ### Review auto categorized data elements
     #### Use 'user_prompt_list.R' to ask the user which rows to edit
