+++ R/user_interactions.R
@@ -155,7 +155,6 @@ user_categorisation_loop <- function(start_v, end_v, table_df, df_prev_exist, df
 #' @return It returns variable text, depending on any_keys.
 #' @keywords internal
 user_prompt <- function(prompt_text, any_keys) {
-
   # prompt text
   if (any_keys == TRUE) {
     response <- ""
