### 2024-12-12 be25741
remove package version comparison from test
