### 2024-12-14 #674 @emilyriederer
Thanks for the submission, @RayStick ! This looks like a great candidate for rOpenSci. I'll begin the search for a handling editor
