### 2024-12-16 #674 @maelle
---

Thank you for your submission! Below are some comments before I start looking for reviewers. Please feel free to ask me any question.

- I think the `DESCRIPTION` and `README` should be more precise about the covered data sources, and their scope (countries, themes) and whether the package can be extended. That's my main concern at the moment.
- Package name: are you sure you want to keep the name camel-case? We recommend lower-case names but at the same time I am aware that camel-case is better for accessibility so this is really just a question.
- I'd recommend removing the "Maintainer" field from `DESCRIPTION` as `Authors@R` already covers it.
- I'd recommend running `desc::desc_normalize()` to normalize `DESCRIPTION`. For instance the dependencies are not alphabetically ordered yet.
- I like how welcoming the contributing guide is!
- Instead of suggesting users start from an empty folder, could the functionality be integrated with for instance `usethis::create_project()`?
- For opening an URL in the browser, you can recommend (or even add in a cli `.run` message) `browseURL()`.
- Why list internal functions in the pkgdown reference index?
- In tests, individual test files do not have to load libraries. https://github.com/aim-rsf/browseMetadata/blob/main/tests/testthat/test-browse_metadata.R
- Instead of `tempdir()` it is safer to use `withr::local_tempfile()` that automatically cleans up after itself. Context: https://testthat.r-lib.org/articles/test-fixtures.html (so you could get rid of https://github.com/aim-rsf/browseMetadata/blob/57b7191b967870160f06dcbec59fd58f36cbbb23/tests/testthat/test-browse_metadata.R#L32) -- that's personal preference though.
- `expect_true(nrow(bar_csv) == 26))` could be `expect_equal(nrow(bar_csv), 26))`. I like reading https://testthat.r-lib.org/reference/index.html#expectations once in a while because I easily forget about shorter expectations (and there are new ones). :sweat_smile:  Actually `expect_equal()` is used in another file. Actually some linters might help you catch such cases: https://lintr.r-lib.org/reference/pkg_testthat_linters.html
- For mocking you might like the new support for mocking in testthat https://testthat.r-lib.org/reference/local_mocked_bindings.html (see also https://blog.r-hub.io/2024/03/21/mocking-new-take/)
- Regarding https://github.com/aim-rsf/browseMetadata/blob/57b7191b967870160f06dcbec59fd58f36cbbb23/R/map_metadata.R#L60 why not simply make `getwd()` the default for the `output_dir` argument? This way it's clearer to users what the default is.
---
