### 2024-12-16 #674 @RayStick
Thanks for these comments @maelle
This morning, I am going to add a rOpenSci review badge to my README, and add a NEWS file
Just to check - is the idea that I respond and implement your suggested changes above before it goes to review?
