### 2024-12-16 4bd102f
Create NEWS.md
