+++ NEWS.md
@@ -1,0 +1,21 @@
+browseMetadata 2.0.2 (2024-12-12)
+=========================
+
+🥳 Submitted this realease to rOpenSci: https://github.com/ropensci/software-review/issues/674
+
+### NEW FEATURES
+
+* Addition of pkgcheck workflow
+* Added more units test to get the code coverage to exceed minimum standard (>75%)
+
+### MINOR IMPROVEMENTS
+
+* added bug fix location to DESCRIPTION
+* added pkgcheck workflow badge to README
+* inidicated internal functions by using @internal tag
+* simplified the pkgdown file using the internal tag
+* updated pull request template
+* added examples to functions that did not have them 
+  
+
+
