### 2024-12-16 #674 @RayStick
Thanks! Most seem pretty do-able to me, and I should have a bit of time today to go through them
