### 2024-12-16 03fc91a
removing  "Maintainer" field as Authors@R covers it
