### 2024-12-16 9d030b4
run desc_normalize()
