### 2024-12-16 d0a3725
make default output_dir more transparent
