### 2024-12-16 04fc2dd
linter improvements
