+++ R/data_manipulation.R
@@ -40,11 +40,11 @@ json_table_to_df <- function(dataset, n) {
 count_empty_desc <- function(table_df, table_colname) {
   table_df["empty"] <- NA
 
-  for (data_v in 1:nrow(table_df)) {
-    if ((nchar(table_df$description[data_v]) == 1) |
-      (table_df$description[data_v] == "Description to follow") |
-      (table_df$description[data_v] == "NA")) {
-      table_df$empty[data_v] <- "Yes"
+  for (data_v in seq_len(nrow(table_df))) {
+    if ((nchar(table_df$description[data_v]) == 1) ||
+        (table_df$description[data_v] == "Description to follow") ||
+        (table_df$description[data_v] == "NA")) {
+        table_df$empty[data_v] <- "Yes"
     } else {
       table_df$empty[data_v] <- "No"
     }
