+++ tests/testthat/test-ref_plot.R
@@ -1,6 +1,4 @@
-# load libraries
-library(testthat) # test_that, expect_equal, expect_true
-library(mockery) # mock, stub
+# libraries: testthat, mockery
 
 # Unit test for ref_plot function
 test_that("ref_plot function works correctly", {
