+++ tests/testthat/test-count_empty_desc.R
@@ -1,6 +1,4 @@
-# load libraries
-library(testthat) # test_that, expect_equal
-library(tidyr)
+# libraries: testthat, tidyr
 
 test_that("count_empty_desc correctly counts empty descriptions", {
   # Sample input data frame
