### 2024-12-16 957e019
directly open html outputs in the browser
