### 2024-12-16 #674 @RayStick
Question about

>  Package name: are you sure you want to keep the name camel-case? We recommend lower-case names but at the same time I am aware that camel-case is better for accessibility so this is really just a question.

So do you mean change to `browsemetadata`? Or could I have `browse-metadata` (I think perhaps not allowed)?
I am open to changing it, but yeah it is a bit of a balance with clarity/readability
Should I see what reviewers think?
