### 2024-12-16 0dc469f
updates based on new checking functions
