+++ tests/testthat/test-map_metadata_convert.R
@@ -1,8 +1,8 @@
-# libraries: testthat
+# libraries: testthat, withr
 
 test_that("map_metadata_convert function outputs files correctly", {
   # Setup
-  temp_dir <- withr::local_tempdir()
+  temp_dir <- local_tempdir()
   file_in <- system.file("outputs/OUTPUT_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-11-27-14-19-55.csv", package = "browseMetadata")
   file_out <- system.file("outputs/L-OUTPUT_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-11-27-14-19-55.csv", package = "browseMetadata")
 
