### 2024-12-16 #674 @RayStick
Is it acceptable to put you as a reviewer on this PR @maelle?
I am nearly done, just 2 of the points to address
Thank you, the feedback is great
