### 2024-12-16 5d869ae
remove mockery, use local_mocked_bindings
