### 2024-12-16 315f2f3
add browseURL to namespace
