+++ NAMESPACE
@@ -34,6 +34,7 @@ importFrom(stats,reorder)
 importFrom(tidyr,complete)
 importFrom(tidyr,pivot_longer)
 importFrom(tools,file_path_sans_ext)
+importFrom(utils,browseURL)
 importFrom(utils,packageVersion)
 importFrom(utils,read.csv)
 importFrom(utils,write.csv)
