### 2024-12-16 4195f57
update tests to use local_mock and handle imports better
