+++ tests/testthat/test-join_outputs.R
@@ -1,5 +1,3 @@
-# libraries: testthat
-
 # Create sample data (only allow it to be different on timestamp, Domain_code and Note)
 session_1 <- data.frame(
   timestamp = c("2024-08-22-13-26-33", "2024-08-22-13-26-33", "2024-08-22-13-26-33"),

[…]
