### 2024-12-17 #674 @RayStick
> I think the DESCRIPTION and README should be more precise about the covered data sources, and their scope (countries, themes) and whether the package can be extended. That's my main concern at the moment.

This is a good point. I had more info in here in earlier versions but I realised that when I have improved the README in other ways, I have cut out lots of this context. To answer you very broadly here:

The covered data sources are any data sources that are listed on the gateway (https://healthdatagateway.org/en) which is a comprehensive collection of UK health data sources.  In theory, if any dataset is exported with their same structural metadata, it can be used in this package! However, they have very recently made improvements to their gateway (released a couple of months ago) - I used to download the metadata files from here (https://maurosandbox.com/hdruk/#/home) but this doesn't seem to be the best place anymore, and I have just noticed that some backwards compatibility with their service is broken. In future, I want my R package to [use an API ](https://github.com/aim-rsf/browseMetadata/issues/117) to retrieve input metedata files but I have not got that working yet, so I have that down as an enhancement as it is not too difficult for a user to download a file.

My suggestion:
- Can we address my PR as it is, but I return to this (important) point above as a separate PR?
- The PR is ready to review but I am unsure why my test coverage is failing - will investigate
