+++ R/map_metadata_convert.R
@@ -23,7 +23,7 @@ map_metadata_convert <- function(output_csv, output_dir) {
     if (grepl(",", output$domain_code[row])) { # Domain_code for this row is a list
       domain_code_list <- output$domain_code[row] # extract Domain_code list
       domain_code_list_split <- unlist(strsplit(domain_code_list, ",")) # split the list
-      for (code in seq_len(domain_code_list_split)) { # for every domain code in list, create a new row
+      for (code in seq_len(length(domain_code_list_split))) { # for every domain code in list, create a new row
         row_to_copy <- output[row, ] # extract row
         row_to_copy$domain_code <- domain_code_list_split[code] # change domain code to single
         output_long[nrow(output_long) + 1, ] <- row_to_copy # copy altered row
