### 2024-12-17 #674 @maelle
> Can we address my PR as it is, but I return to this (important) point above as a separate PR?

Yes absolutely. However it should be tackled before I invite reviewers, because that's crucial information.

I'll review the current PR now!
