### 2024-12-17 #674 @RayStick
@maelle thank you for your PR review, my package is improving already
Can we move the request to reviewer's until Jan? (I am back to work on 6th Jan)
This week I hope to:
1. Work on implementing the `utils:menu()` as you suggested, to simplify the code base
2. Work on updating the README for more package context (this doesn't sound like a big job but unfortunately I think my code might have to change as I noticed the input file structure on the latest metadata catalogue has changed!)
