### 2024-12-17 22673f7
combined browse into map, did some renaming
