+++ R/plotting.R
@@ -1,10 +1,10 @@
 #' Internal: ref_plot
 #'
-#' This function is called within the map_metadata function. \cr \cr
+#' This function is called within the map.R function. \cr \cr
 #' It plots a reference table to guide the user in their categorisation of domains. \cr \cr
 #' This reference table is based on the user inputted domains and the default domains provided by this package.  \cr \cr
 #' @param domains The output of load_data
-#' @return A reference table that appears in the Plots tab. A list of 2 containing the derivatives for this plot, used later in map_metadata'
+#' @return A reference table that appears in the Plots tab. A list of 2 containing the derivatives for this plot, used later in map.R'
 #' @keywords internal
 #' @importFrom gridExtra tableGrob grid.arrange
 #' @importFrom graphics plot.new
@@ -22,7 +22,7 @@ ref_plot <- function(domains) {
 
 #' Internal: end_plot
 #'
-#' This function is called within the map_metadata function. \cr \cr
+#' This function is called within the map.R function. \cr \cr
 #' A summary plot is created that includes the domain code reference table and counts of domain code categorisations \cr \cr
 #'
 #' @param df The Output dataframe with all the domain categorisations
