+++ tests/testthat/test-data_load.R
@@ -8,34 +8,34 @@ json <- get("json_metadata")
 look_up <- get("look_up")
 domains <- get("domain_list")
 
-test_that("load_data runs in demo mode when both json_file and domain_file are NULL", {
-  result <- load_data(NULL, NULL, NULL)
+test_that("data_load runs in demo mode when both json_file and domain_file are NULL", {
+  result <- data_load(NULL, NULL, NULL)
   expect_true(result$demo_mode)
   expect_equal(result$meta_json, json)
   expect_equal(result$domains, domains)
   expect_equal(result$domain_list_desc, "DemoList")
 })
 
-test_that("load_data throws error if only one of json_file or domain_file is NULL", {
-  expect_error(load_data(json_file, NULL, NULL))
-  expect_error(load_data(NULL, domains_file, NULL))
+test_that("data_load throws error if only one of json_file or domain_file is NULL", {
+  expect_error(data_load(json_file, NULL, NULL))
+  expect_error(data_load(NULL, domains_file, NULL))
 })
 
-test_that("load_data reads user-specified files correctly", {
-  result <- load_data(json_file, domains_file, NULL)
+test_that("data_load reads user-specified files correctly", {
+  result <- data_load(json_file, domains_file, NULL)
   expect_false(result$demo_mode)
   expect_true(is.list(result$meta_json))
   expect_true(is.data.frame(result$domains))
   expect_equal(result$domain_list_desc, tools::file_path_sans_ext(basename(domains_file)))
 })
 
-test_that("load_data uses default look-up table when look_up_file is NULL", {
-  result <- load_data(json_file, domains_file, NULL)
+test_that("data_load uses default look-up table when look_up_file is NULL", {
+  result <- data_load(json_file, domains_file, NULL)
   expect_equal(result$lookup, look_up)
 })
 
-test_that("load_data reads user-specified look-up table correctly", {
-  result <- load_data(json_file, domains_file, look_up_file)
+test_that("data_load reads user-specified look-up table correctly", {
+  result <- data_load(json_file, domains_file, look_up_file)
   expect_true(is.data.frame(result$lookup))
   expect_equal(nrow(result$lookup), nrow(utils::read.csv(look_up_file)))
 })
