+++ tests/testthat/test-outputs_join.R
@@ -51,9 +51,9 @@ expected_output_ses1_ses1 <- data.frame(
 )
 
 # Write the test
-test_that("join_outputs works correctly", {
-  result <- join_outputs(session_1, session_1)
+test_that("outputs_join works correctly", {
+  result <- outputs_join(session_1, session_1)
   expect_equal(result, expected_output_ses1_ses1)
-  result <- join_outputs(session_1, session_2)
+  result <- outputs_join(session_1, session_2)
   expect_equal(result, expected_output_ses1_ses2)
 })
