### 2024-12-18 3484f58
allow metadata_map to only produce plots
