### 2024-12-19 687c204
moving from json to csv (not all unit tests pass yet)
