### 2024-12-19 4174c77
remove custom functions, update unit tests
