+++ tests/testthat/test-metadata_map.R
@@ -2,29 +2,32 @@ test_that("metadata_map function works correctly with user input", {
   # Setup
   temp_dir <- withr::local_tempdir()
 
-  demo_log_output <- system.file("outputs/LOG_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-11-27-14-19-55.csv", package = "browseMetadata")
-  demo_output <- system.file("outputs/OUTPUT_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-11-27-14-19-55.csv", package = "browseMetadata")
+  demo_log_output <- system.file("outputs/MAPPING_LOG_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv", package = "browseMetadata")
+  demo_output <- system.file("outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv", package = "browseMetadata")
+  demo_bar <- system.file("outputs/BAR_360_NationalCommunityChildHealthDatabase(NCCHD)_2024-12-19-14-11-55.csv", package = "browseMetadata")
+  # IMPROVE - also test MAPPING_PLOT_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.png
+  # IMPROVE - also test BAR_360_NationalCommunityChildHealthDatabase(NCCHD)_2024-12-19-14-11-55.html
 
   # Mock functions
   local_mocked_bindings(
     readline = function(prompt) {
       response <- switch(prompt,
-        "Press any key to continue " = "", # line 109 in map.R
-        "Optional free text note about this table (or press enter to continue): " = "demo run" # line 109 in map.R
+        "Press 'Esc' key to finish here, or press any other key to continue with mapping variables" = "1", # line 108
+        "Enter your initials: " = "rs", # line 127
+        "Optional free text note about this table (or press enter to continue): " = "demo run" # line 147
       )
     }
   )
 
-  local_mocked_bindings(
-    user_prompt = function(prompt_text, any_keys) {
-      if (prompt_text == "Enter your initials: ") { # line 93 in map.R
-        response <- "demo"
-      } else if (prompt_text == "Would you like to review your categorisations? (y/n): ") { # line 238 in map.R
-        response <- "n"
-      }
-      return(response)
-    }
-  )
+    local_mocked_bindings(
+      menu = function(choices, graphics = FALSE, title = NULL) {
+        response <- switch(title,
+                           "Enter the table numbers you want to process:" = 4,
+                           "\nWould you like to review your categorisations?" = 0 # IMPROVE - could test this with not 0
+        )
+        return(response)
+      })
+
 
   local_mocked_bindings(
     user_categorisation_loop = function(start_v, end_v, table_df, df_prev_exist, df_prev, lookup, df_plots, output_df) {
@@ -36,20 +39,20 @@ test_that("metadata_map function works correctly with user input", {
   )
 
   local_mocked_bindings(
-    user_prompt_list = function(prompt_text, list_allowed, empty_allowed) {
-      if (grepl("Enter row numbers for those you want to edit:", prompt_text)) {
-        return(list()) # return an empty list for the auto categorised data elements prompt
-      }
-      return(list(2)) # return a list with '2' which means choosing the table CHILD
-    }
-  )
+    select.list = function(choices, preselect = NULL, multiple = FALSE, title = NULL, graphics = getOption("menu.graphics"), inline = getOption("menu.graphics")) {
+      response <- switch(title,
[…]
