+++ R/metadata_map.R
@@ -1,4 +1,6 @@
 readline <- NULL
+menu <- NULL
+select.list <- NULL
 
 #' metadata_map
 #'
@@ -45,7 +47,7 @@ readline <- NULL
 #' @export
 #' @importFrom dplyr %>% filter
 #' @importFrom cli cli_h1 cli_alert_info cli_alert_success
-#' @importFrom utils packageVersion write.csv browseURL
+#' @importFrom utils packageVersion write.csv browseURL menu select.list
 #' @importFrom ggplot2 ggsave
 #' @importFrom htmlwidgets saveWidget
 
@@ -123,222 +125,168 @@ metadata_map <- function(
     stop()
   }
 
-  ## Use 'user_prompt.R' to get user initials for the log file
-  user_initials <- user_prompt(
-    prompt_text = "Enter your initials: ", any_keys = TRUE
-  )
+  ## Get user initials for the log file
+  user_initials <- readline(prompt = "Enter your initials: ")
 
-  ## WHICH TABLES FROM THE DATASET?
+  ## CHOOSE TABLE TO PROCESS
 
-  ### Use 'user_prompt_list.R' to ask user which tables to process
+  chosen_table_n <- menu(levels(dataset$Section),title = 'Enter the table number you want to process:')
+  table_name <- levels(dataset$Section)[chosen_table_n]
   cat("\n")
-  for (dc in 1:n_tables) {
-    table_name <- levels(dataset$Section)[dc]
-    cat(dc, ". ", table_name, "\n", sep = "")
+  cli_alert_info("Processing Table {chosen_table_n} of {n_tables} ({table_name})")
+  cat("\n")
+
+  #### Use 'output_copy.R' to copy from previous output(s) if they exist
+  if (table_copy == TRUE) {
+    copy_prev <- output_copy(dataset_name, output_dir)
+    df_prev_exist <- copy_prev$df_prev_exist
+    df_prev <- copy_prev$df_prev
+  } else {
+    df_prev_exist <- FALSE
   }
 
-  n_tables_process <- user_prompt_list(
-    prompt_text =
-      paste("Enter table numbers you want to process (one table number on each line)."),
-    list_allowed = seq(from = 1, to = n_tables, by = 1),
-    empty_allowed = FALSE
-  )
+  table_note <- readline(paste(
+    "Optional free text note about this table",
+    "(or press enter to continue): "
+  ))
 
-  ## PROCESS EACH CHOSEN TABLE
+  ####  Extract table from metadata
+  table_df <- dataset %>% filter(Section == levels(dataset$Section)[chosen_table_n])
 
-  ### Extract each Table
-  for (dc in unique(n_tables_process)) {
-    table_name <- levels(dataset$Section)[dc]
-    cat("\n")
-    cli_alert_info("Processing Table {dc} of {n_tables} ({table_name})")
-    cat("\n")
+  #### If demo, only process the first 20 elements
+  if (data$demo_mode == TRUE) {
+    start_v <- 1
+    end_v <- min(20, nrow(table_df))
+  } else {
+    start_v <- 1
+    end_v <- nrow(table_df)
+  }
 
-    #### Use 'output_copy.R' to copy from previous output(s) if they exist
-    if (table_copy == TRUE) {
-      copy_prev <- output_copy(dataset_name, output_dir)
-      df_prev_exist <- copy_prev$df_prev_exist
-      df_prev <- copy_prev$df_prev
-    } else {
-      df_prev_exist <- FALSE
+  #### Use 'user_categorisation_loop.R' to copy or request from user
+
+  output_df <- user_categorisation_loop(start_v,
+                                        end_v,
+                                        table_df,
+                                        df_prev_exist,
+                                        df_prev,
+                                        lookup = data$lookup,
[…]
