+++ NAMESPACE
@@ -30,6 +30,8 @@ importFrom(stats,reorder)
 importFrom(tidyr,pivot_longer)
 importFrom(tools,file_path_sans_ext)
 importFrom(utils,browseURL)
+importFrom(utils,menu)
 importFrom(utils,packageVersion)
 importFrom(utils,read.csv)
+importFrom(utils,select.list)
 importFrom(utils,write.csv)

[…]
