+++ README.md
@@ -1,24 +1,22 @@
-# mmetadata (map metadata)
+# mapmetadata
 
  *Mapping from variables to concepts*
 
-<a href="https://aim-rsf.github.io/browseMetadata/"><img src="man/figures/logo.png" align="right" height="120" alt="browseMetadata website" /></a>
+<a href="https://aim-rsf.github.io/mapmetadata/"><img src="man/figures/logo.png" align="right" height="120" alt="mapmetadata website" /></a>
 
 <!-- badges: start -->
 [![All Contributors](https://img.shields.io/badge/all_contributors-4-orange.svg?style=flat-square)](#contributors-) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10581500.svg)](https://doi.org/10.5281/zenodo.10581500)
 
-[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) [![R-CMD-check](https://github.com/aim-rsf/browseMetadata/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/aim-rsf/browseMetadata/actions/workflows/R-CMD-check.yaml)
-[![pkgcheck](https://github.com/aim-rsf/browseMetadata/workflows/pkgcheck/badge.svg)](https://github.com/aim-rsf/browseMetadata/actions?query=workflow%3Apkgcheck) [![codecov](https://codecov.io/gh/aim-rsf/browseMetadata/graph/badge.svg?token=59S2QVG7CQ)](https://codecov.io/gh/aim-rsf/browseMetadata) [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/674_status.svg)](https://github.com/ropensci/software-review/issues/674)
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) [![R-CMD-check](https://github.com/aim-rsf/mapmetadata/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/aim-rsf/mapmetadata/actions/workflows/R-CMD-check.yaml)
+[![pkgcheck](https://github.com/aim-rsf/mapmetadata/workflows/pkgcheck/badge.svg)](https://github.com/aim-rsf/mapmetadata/actions?query=workflow%3Apkgcheck) [![codecov](https://codecov.io/gh/aim-rsf/mapmetadata/graph/badge.svg?token=59S2QVG7CQ)](https://codecov.io/gh/aim-rsf/mapmetadata) [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/674_status.svg)](https://github.com/ropensci/software-review/issues/674)
 <!-- badges: end -->
 
 # Table of Contents
 
-1. [What is the `mmetaddata` package?](#what-is-the-mmetaddata-package)
-2. [Getting started with `mmetaddata`](#getting-started-with-mmetaddata)
-   - [Installation and set-up](#installation-and-set-up)
-   - [Demo (using the `R Studio` IDE)](#demo-using-the-r-studio-ide)
-     - [`browseMetadata()`](#browsemetadata-1)
-     - [`mapMetadata()`](#mapmetadata)
+1. [What is the `mapmetadata` package?](#what-is-the-mapmetadata-package)
+2. [Getting started with `mapmetadata`](#getting-started-with-mapmetadata)
+ - 2a. [Installation and set-up](#installation-and-set-up)
+ - 2b. [Demo (using the `R Studio` IDE)](#demo-using-the-r-studio-ide)
[…]
