+++ vignettes/mapmetadata.Rmd
@@ -1,0 +1,209 @@
+---
+title: "mapmetadata tutorial"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{mapmetadata tutorial}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+For installation, set-up and basic usage refer to the package [README.md](https://aim-rsf.github.io/mapmetadata/index.html) file. In this tutorial, you will find more context on package inputs, how to interact with package functions and interpret package outputs.
+
+# Preparing for mapping
+
+<img src="https://raw.githubusercontent.com/aim-rsf/mapmetadata/main/inst/outputs/BAR_360_NationalCommunityChildHealthDatabase(NCCHD)_2024-12-19-18-07-22.png" alt="example bar plot showing number of variables for each table alongside counts of whether variables have missing descriptions"/>
+
+This bar plot is produced automatically when you run the `metadata_map` function. The bar plot can help you understand the scope of the dataset, but reference the HDRUK Gateway page for the fuller context. For instance, table descriptions are not included in these structural metadata files but they are included on the gateway:
+
+**For dataset NCCHD, used in the demo, the structural metadata was downloaded here:**
+
+<img src="https://raw.githubusercontent.com/aim-rsf/mapmetadata/main/inst/outputs/inst/outputs/NCCHD_gateway_download.png" alt="screenshot of HRDRUK gateway showing location to download structural metadata"/>
+
+**View descriptions of each table to give you more context on their contents and how to use them:**
+
+<img src="https://raw.githubusercontent.com/aim-rsf/mapmetadata/main/inst/outputs/inst/outputs/NCCHD_gateway_table.png" alt="screenshot of HRDRUK gateway showing table descriptions"/>
+
+The bar plot shows us there are 13 tables in the dataset. The height of the bar indicates the number of variables in that table:
+
+-   The ones with lots of variables (e.g. CHILD_TRUST) will take you longer to process
+-   Some tables (e.g. CHE_HEALTHYCHILDWALESPROGRAMME) have a lot of empty descriptions. An empty description means that this variable will only have a label and a data type.
+
+It is important to note that this plot is only summarising *variable* level metadata i.e. a description of what the variable is. Some variables also require *value* level metadata i.e. what does each value correspond to, 1 = Yes, 2 = No, 3 = Unknown. This *value* level metadata can sometimes be found in lookup tables, if it is not provided within the *variable* level description.
+
+# Mapping
+
+Use the bar plot and the HDRUK Gateway to guide your mapping choices. The main functionality of this package is to aid a researcher in mapping variables from health datasets onto their research domains (concepts/latent variables).
+
+You will see this in the R console:
+
+`Press 'Esc' key to finish here, or press any other key to continue with mapping variables`
+
+Pressing any key will move you to the mapping phase. Demo mode (run by `metadata_map()`) only processes the fist 20 data elements (variables) in the selected table.
+
+```         
+Enter your initials: rs
+
+Enter the table number you want to process: 
+
+ 1: BLOOD_TEST    
+ 2: BREAST_FEEDING                   
+ 3: CHE_HEALTHYCHILDWALESPROGRAMME   
+ 4: CHILD                   
+ 5: CHILD_BIRTHS  
+ 6: CHILD_MEASUREMENT_PROGRAM        
+ 7: CHILD_TRUST                      
+ 8: EXAM                    
+ 9: IMM           
+ 10: PATH_BLOOD_TESTS                
[…]
