+++ _pkgdown.yml
@@ -22,5 +22,6 @@ navbar:
 reference:
 - title: "User functions"
   contents:
-  - starts_with("map_metadata")
-  - browse_metadata
+  - metadata_map
+  - map_compare
+  - map_convert 
