+++ tests/testthat/test-map_compare.R
@@ -2,11 +2,11 @@ test_that("map_compare function works correctly with user input", {
   # Setup
   temp_dir <- withr::local_tempdir()
 
-  demo_session_dir <- system.file("outputs", package = "browseMetadata")
+  demo_session_dir <- system.file("outputs", package = "mapmetadata")
   demo_session1_base <- "360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55"
   demo_session2_base <- "360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45"
-  demo_metadata_file <- system.file("inputs", "360_National Community Child Health Database (NCCHD)_Structural_Metadata.csv", package = "browseMetadata")
-  demo_domain_file <- system.file("inputs", "domain_list_demo.csv", package = "browseMetadata")
+  demo_metadata_file <- system.file("inputs", "360_National Community Child Health Database (NCCHD)_Structural_Metadata.csv", package = "mapmetadata")
+  demo_domain_file <- system.file("inputs", "domain_list_demo.csv", package = "mapmetadata")
 
   # mock concensus_on_mismatch
   local_mocked_bindings(
