+++ man/mapmetadata-package.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/browseMetadata-package.R
+% Please edit documentation in R/mapmetadata-package.R
 \docType{package}
-\name{browseMetadata-package}
-\alias{browseMetadata}
-\alias{browseMetadata-package}
-\title{browseMetadata: Map health metadata onto predefined research domains}
+\name{mapmetadata-package}
+\alias{mapmetadata}
+\alias{mapmetadata-package}
+\title{mapmetadata: Map health metadata onto predefined research domains}
 \description{
 \if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}
 
@@ -13,8 +13,8 @@ Visualise and categorise publicly available metadata from health datasets. By in
 \seealso{
 Useful links:
 \itemize{
-  \item \url{https://aim-rsf.github.io/browseMetadata/}
-  \item Report bugs at \url{https://github.com/aim-rsf/browseMetadata/issues}
+  \item \url{https://aim-rsf.github.io/mapmetadata/}
+  \item Report bugs at \url{https://github.com/aim-rsf/mapmetadata/issues}
 }
 
 }
