+++ man/map_convert.Rd
@@ -21,7 +21,7 @@ This 'L-OUTPUT_' may be more useful when using these csv files in later analyses
 }
 \examples{
 # Locate file path and file name for the example files in the package
-demo_output_dir <- system.file("outputs", package = "browseMetadata")
+demo_output_dir <- system.file("outputs", package = "mapmetadata")
 demo_output_csv <- "MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv"
 
 # Run the function
