+++ R/empty_count.R
@@ -6,7 +6,7 @@
 #' @return Returns a long dataframe with 3 columns: Empty (No, Yes), Table (table name), N_Variables (count).
 #' @keywords internal
 #' @importFrom tidyr pivot_longer
-#' @importFrom dplyr %>% group_by summarise mutate select arrange
+#' @importFrom dplyr %>% group_by summarise mutate select arrange n
 
 empty_count <- function(dataframe) {
 
