+++ NAMESPACE
@@ -16,7 +16,10 @@ importFrom(dplyr,count)
 importFrom(dplyr,distinct)
 importFrom(dplyr,filter)
 importFrom(dplyr,group_by)
+importFrom(dplyr,join_by)
+importFrom(dplyr,left_join)
 importFrom(dplyr,mutate)
+importFrom(dplyr,n)
 importFrom(dplyr,select)
 importFrom(dplyr,summarise)
 importFrom(ggplot2,ggsave)
