+++ R/metadata_map.R
@@ -58,7 +58,6 @@ metadata_map <- function(
     output_dir = getwd(),
     table_copy = TRUE,
     long_output = TRUE) {
-
   timestamp_now_fname <- format(Sys.time(), "%Y-%m-%d-%H-%M-%S")
   timestamp_now <- format(Sys.time(), "%Y-%m-%d %H:%M:%S")
 
@@ -87,11 +86,13 @@ metadata_map <- function(
   empty_count_df <- empty_count(dataset)
 
   ## Use 'empty_plot.R' to create bar plot then save it
-  barplot_html <- empty_plot(empty_count_df,n_tables)
+  barplot_html <- empty_plot(empty_count_df, n_tables)
   original_wd <- getwd()
   setwd(output_dir) # saveWidget has a bug with paths & saving
-  base_fname <-  paste0("BAR_", gsub(" ", "", dataset_name),"_",
-                        timestamp_now_fname)
+  base_fname <- paste0(
+    "BAR_", gsub(" ", "", dataset_name), "_",
+    timestamp_now_fname
+  )
   bar_fname <- paste0(base_fname, ".html")
   saveWidget(widget = barplot_html, file = bar_fname, selfcontained = TRUE)
   bar_data_fname <- paste0(base_fname, ".csv")
@@ -130,7 +131,7 @@ metadata_map <- function(
 
   ## CHOOSE TABLE TO PROCESS
 
-  chosen_table_n <- menu(levels(dataset$Section),title = 'Enter the table number you want to process:')
+  chosen_table_n <- menu(levels(dataset$Section), title = "Enter the table number you want to process:")
   table_name <- levels(dataset$Section)[chosen_table_n]
   cat("\n")
   cli_alert_info("Processing Table {chosen_table_n} of {n_tables} ({table_name})")
@@ -165,13 +166,14 @@ metadata_map <- function(
   #### Use 'user_categorisation_loop.R' to copy or request from user
 
   output_df <- user_categorisation_loop(start_v,
-                                        end_v,
-                                        table_df,
-                                        df_prev_exist,
-                                        df_prev,
-                                        lookup = data$lookup,
-                                        df_plots,
-                                        output_df)
+    end_v,
+    table_df,
+    df_prev_exist,
+    df_prev,
+    lookup = data$lookup,
+    df_plots,
+    output_df
+  )
 
   output_df$timestamp <- timestamp_now
   output_df$table <- table_name
@@ -179,17 +181,18 @@ metadata_map <- function(
 
   #### Review auto categorized data elements
   cat("\n")
-  cli_alert_info('These are the auto categorised data elements:')
+  cli_alert_info("These are the auto categorised data elements:")
   cat("\n")
   output_auto <- subset(output_df, note == "AUTO CATEGORISED")
   output_auto <- output_auto[, c("data_element", "domain_code", "note")]
-  print(output_auto,row.names = FALSE)
+  print(output_auto, row.names = FALSE)
 
   auto_elements <- output_df$data_element[output_df$note == "AUTO CATEGORISED"]
 
   auto_row_names <- select.list(auto_elements,
-                                 multiple = TRUE,
-                                 title = "\nSelect those you want to manually edit:")
+    multiple = TRUE,
+    title = "\nSelect those you want to manually edit:"
+  )
 
   auto_row <- which(output_df$data_element %in% auto_row_names)
 
@@ -210,20 +213,21 @@ metadata_map <- function(
 
   ### Review user categorized data elements (optional)
   cat("\n")
[…]
