+++ tests/testthat/test-empty_count.R
@@ -1,10 +1,10 @@
 test_that("empty_count correctly counts empty descriptions", {
   # Sample input data frame
   table_df <- data.frame(
-    Section = c("Table1","Table1","Table1","Table2","Table2","Table2"),
-    Column.name = c("var1", "var2", "var3", "var4", "var5","var6"),
+    Section = c("Table1", "Table1", "Table1", "Table2", "Table2", "Table2"),
+    Column.name = c("var1", "var2", "var3", "var4", "var5", "var6"),
     Column.description = c("Description to follow", "Valid description", "NA", "Another valid description", " ", "-"),
-    Data.type = c("VARCHAR", "BIGINT", "DATE", "TIMESTAMP", "VARCHAR","VARCHAR"),
+    Data.type = c("VARCHAR", "BIGINT", "DATE", "TIMESTAMP", "VARCHAR", "VARCHAR"),
     stringsAsFactors = FALSE
   )
 
@@ -12,9 +12,9 @@ test_that("empty_count correctly counts empty descriptions", {
 
   # Expected output data frame
   expected_output <- tidyr::tibble(
-    Table = c("Table1","Table1","Table2","Table2"),
-    Empty = c("No", "Yes","No", "Yes"),
-    N_Variables = c(1,2,1,2)
+    Table = c("Table1", "Table1", "Table2", "Table2"),
+    Empty = c("No", "Yes", "No", "Yes"),
+    N_Variables = c(1, 2, 1, 2)
   )
 
   expected_output$Table <- as.factor(expected_output$Table)
