+++ pkgdown.yml
@@ -2,9 +2,9 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
-  browseMetadata: browseMetadata.html
   HealthMetadata: HealthMetadata.html
-last_built: 2024-12-17T11:22Z
+  mapmetadata: mapmetadata.html
+last_built: 2024-12-20T11:57Z
 urls:
-  reference: https://aim-rsf.github.io/browseMetadata/reference
-  article: https://aim-rsf.github.io/browseMetadata/articles
+  reference: https://aim-rsf.github.io/mapmetadata/reference
+  article: https://aim-rsf.github.io/mapmetadata/articles
