+++ tests/testthat/test-data_load.R
@@ -1,5 +1,5 @@
 # Define file paths to demo data relative to the package directory
-csv_file <- system.file("inputs/360_National Community Child Health Database (NCCHD)_Structural_Metadata.csv", package = "mapmetadata")
+csv_file <- system.file("inputs/360_NationalCommunityChildHealthDatabase(NCCHD)_Structural_Metadata.csv", package = "mapmetadata")
 look_up_file <- system.file("inputs/look_up.csv", package = "mapmetadata")
 domains_file <- system.file("inputs/domain_list_demo.csv", package = "mapmetadata")
 
