### 2025-01-06 ea33821
update desc to directly mention HDRUK
