### 2025-01-06 032bb1b
add dataset title to BAR plot
