+++ vignettes/mapmetadata.Rmd
@@ -18,7 +18,7 @@ For installation, set-up and basic usage refer to the package [README.md](https:
 
 # Preparing for mapping
 
-<img src="https://raw.githubusercontent.com/aim-rsf/mapmetadata/main/inst/outputs/BAR_360_NationalCommunityChildHealthDatabase(NCCHD)_2024-12-19-18-07-22.png" alt="example bar plot showing number of variables for each table alongside counts of whether variables have missing descriptions"/>
+<img src="https://raw.githubusercontent.com/aim-rsf/mapmetadata/main/inst/outputs/BAR_360_NationalCommunityChildHealthDatabase(NCCHD)_2024-12-19-14-11-55.png" alt="example bar plot showing number of variables for each table alongside counts of whether variables have missing descriptions"/>
 
 This bar plot is produced automatically when you run the `metadata_map` function with the demo metadata file, which contains metadata about the **National Community Child Health Database (NCCHD).**
 
