### 2025-01-06 14e0592
update citation files
