+++ CITATION.cff
@@ -1,10 +1,10 @@
 cff-version: 1.2.0
 message: "If you use this package, please cite it as below."
-title: "browseMetadata"
-version: 2.0.2
-date-released: 2024-11-26
+title: "mapmetadata"
+version: 3.0.0
+date-released: 2025-01-06
 authors:
   - family-names: "Rachael"
     given-names: "Stickland"
     orcid: "https://orcid.org/0000-0003-3398-4272"
-repository-code: "https://github.com/aim-rsf/browseMetadata"
+repository-code: "https://github.com/aim-rsf/mapmetadata"
