### 2025-01-06 329bd92
Update NEWS.md
