### 2025-01-06 #674 @RayStick
@maelle this package is ready for review again =D

I have made lots of changes and these are summarised in the [NEWS file](https://github.com/aim-rsf/mapmetadata/blob/main/NEWS.md) and [release notes ](https://github.com/aim-rsf/mapmetadata/releases/tag/v3.0.0) - let me know if you have any remaining questions before sending for review, and I am happy to clarify and/or make any more changes that you think about necessary.

Sorry if this means another round of editor checks for you! The good thing is that the code base is simpler now, and the documentation clearer.
