### 2025-01-07 #674 @RayStick
> I see some functions changed names, old functions were simply removed. I think it's fine at this development stage and especially since the package changed names too. For later in development, I'd like to mention https://lifecycle.r-lib.org/articles/communicate.html in case you don't know about it.

I did not know about this, thank you. I will definitely bear this in mind for later development. I did not expect to make a major release but code refactoring combined with the metadata input files changing made it seem necessary. I understand function history needs to be tracked in a more sophisticated way down the line! Thanks for the resource.

> The README.md has a table of contents, but GitHub also provides an automatic table of contents: https://github.blog/changelog/2021-04-13-table-of-contents-support-in-markdown-files/ On the one hand, removing the table of contents from the Markdown file would reduce clutter and get the reader to the important info sooner. On the other hand, maybe not everyone knows about the automatic table of contents. (and on the pkgdown homepage there's no table of contents, I think because READMEs are often short?)

This is a good point. I think this is a bit of a legacy TOC that seemed more necessary when the README was longer. I am going to remove it!

> For installation, I'd recommend documenting pak instead of devtools as seen in https://usethis.r-lib.org/#installation

Can I put this as an 'enhancement' Issue and tackle this down the line/during review process?

Thanks!
