### 2025-01-17 #674 @ymansiaux
Hi @maelle and @RayStick , please find below my review.

- [x]  As the reviewer I confirm that there are no [[conflicts of interest](https://devguide.ropensci.org/policies.html#coi)](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

Please see below my review comments. I think that some modifications should be made, for example in the README or function examples.

Thank you very much for sharing this work and for your high-quality development.

I particularly appreciate the attention you've paid to documentation and unit testing. I’ve personally learned a lot, especially about mocking, which I hadn’t explored much until now.

Below, please find my detailed comments and suggestions:

**Function and argument naming**

To align with the naming of the other exported functions, `map_compare()` and `map_convert()`, I would suggest renaming `metadata_map()` to `map_metadata()`. While I understand you may have concerns about potential confusion with the package name, I believe this change would harmonize the naming scheme, enhance clarity, and facilitate autocomplete for users.

For the `map_convert()` function, I recommend renaming the argument `output_csv` to `csv_to_convert` to make its purpose clearer. Additionally, introducing a new argument, `csv_to_convert_dir`, could address an issue I’ve identified in the provided function example (see the **Examples** section below). If you implement this change, the `output_dir` argument could represent the directory where the generated CSV file is stored (defaulting to `csv_to_convert_dir` if not specified).

**Input checking**

I suggest adding explicit input validation to your exported functions. This would enhance the user experience by providing clear and informative error messages or warnings.

- For `map_compare()`, consider verifying that `session_dir` exists and that the CSV files specified in the user input are present.
- For `map_convert()`, you might want to check that both `output_csv` and `output_dir` exist.
- For `metadata_map()` and the other functions, validating file existence, expected dimensions, and column names (if any) could further improve robustness.

**README**

The demo example in the README feels somewhat extensive and might benefit from being simplified to improve accessibility. Currently, interacting with up to 20 variables requires substantial effort, which could deter users from running the full demo.

For sections involving custom inputs (metadata input, domain list input, and custom lookup), I recommend moving these to the vignette and referencing it in the README. For example:

> "If you wish to use custom metadata, domain lists, or lookup tables, please refer to vignette XXX, section YYY."
>

Additionally, examples in the README should be copy-pasteable and work seamlessly without requiring additional setup. For instance, the following code snippet requires downloading a file.

```r
new_csv_file <- "path/your_new_csv.csv"
demo_domains_file <- system.file("inputs/domain_list_demo.csv", package = "mapmetadata")

metadata_map(csv_file = new_csv_file, domain_file = demo_domains_file)
```

This is why I would present only the demo example in the README.

Providing examples for custom domain list inputs and custom lookup tables would also be beneficial. For these examples, linking to demo files in `inst/inputs` could clarify the expected file formats. I guess an explanation about the location of the files used by default and the way to access them programmatically with `system.file()` would be useful to your users (If this function is probably familiar to R packages developpers, it might not be straightforward for package “users” only).

For non-exported functions, you might want to prevent `.Rd` files from being generated in the `man/` folder. Adding the `#' @noRd` tag to these functions' documentation can achieve this. For more details, see [[the rOpenSci dev guide](https://devguide.ropensci.org/pkg_building.html#roxygen-2-use)](https://devguide.ropensci.org/pkg_building.html#roxygen-2-use).

**Testing**

In the `tests/testthat.R` file, it seems unnecessary to load the `{mockery}` and `{withr}` packages explicitly, as you are already using explicit namespacing.

A note regarding `{mockery}`: its README states that the package is superseded and recommends using `testthat::local_mocked_bindings()` instead. More details can be found [[here](https://github.com/r-lib/mockery)](https://github.com/r-lib/mockery).

**Examples**

Some examples are enclosed in `dontrun` tags, which may pose an issue if you plan to submit the package to CRAN.

- For `map_compare()`, consider using a temporary directory to avoid creating files in the user’s workspace or package’s installation directory:

```r
# Locate file paths for the example files in the package
demo_session_dir <- system.file("outputs", package = "mapmetadata")
demo_session1_base <- "360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55"
demo_session2_base <- "360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45"
demo_metadata_file <- system.file("inputs", "360_NationalCommunityChildHealthDatabase(NCCHD)_Structural_Metadata.csv", package = "mapmetadata")
demo_domain_file <- system.file("inputs", "domain_list_demo.csv", package = "mapmetadata")
output_dir <- tempdir()
message("Output directory: ", output_dir)

# Run the function - requires user interaction
map_compare(
  session_dir = demo_session_dir,
  session1_base = demo_session1_base,
  session2_base = demo_session2_base,
  metadata_file = demo_metadata_file,
  domain_file = demo_domain_file,
  output_dir = output_dir
)

```

Returning the path of the output file and including a `browseURL()` call to open the file in the example could enhance usability. This would allow us to delete the temp directory once the generated file has been opened.

—

The example for `map_convert()` writes to the package’s installation directory, which is problematic. Modifying this to use temporary directories would improve clarity and compliance. If you implement my earlier suggestion to add a `csv_to_convert_dir` argument, this example could be significantly simplified. Moreover, I think a message telling the user the path of the output could be nice, like you did with `map_compare()`. I think this could be nice that this function returns the path of the output file so we could open it in the example with a `browseURL()` call.

—

For `metadata_map()`, since it requires user interaction, you should enclose it in an `if(interactive())` block:

```r
#' @examples
# Demo run requires no function inputs but requires user interaction
if(interactive()) {
	metadata_map()
}
```

**file path construction**

In your file-handling functions, you are constructing paths using `paste()` or `paste0()`. To ensure platform independence, I recommend replacing these with `file.path()`. For example, in `map_compare()`:

You could replace

```r
  csv_1a <- read.csv(paste0(session_dir, "/", "MAPPING_LOG_", session1_base, ".csv"))
```

with

```r
  csv_1a <- read.csv(file.path(session_dir, paste0("MAPPING_LOG_", session1_base, ".csv")))
```

This ensures compatibility across operating systems.

**devtools::check()**

Here is my session info

```r
R version 4.4.1 (2024-06-14 ucrt)
Platform: x86_64-w64-mingw32/x64
Running under: Windows 11 x64 (build 22631)

[1] LC_COLLATE=French_France.utf8  LC_CTYPE=French_France.utf8
[3] LC_MONETARY=French_France.utf8 LC_NUMERIC=C
[5] LC_TIME=French_France.utf8

time zone: Europe/Paris
tzcode source: internal

[1] mapmetadata_3.0.0 testthat_3.2.2

 [1] gtable_0.3.6      jsonlite_1.8.9    dplyr_1.1.4
 [4] compiler_4.4.1    brio_1.1.5        tidyselect_1.2.1
 [7] gridExtra_2.3     tidyr_1.3.1       scales_1.3.0
[10] yaml_2.3.10       fastmap_1.2.0     ggplot2_3.5.1
[13] R6_2.5.1          generics_0.1.3    knitr_1.49
[16] htmlwidgets_1.6.4 tibble_3.2.1      desc_1.4.3
[19] munsell_0.5.1     rprojroot_2.0.4   pillar_1.10.1
[22] rlang_1.1.4.9000  xfun_0.50         fs_1.6.5
[25] lazyeval_0.2.2    pkgload_1.4.0     viridisLite_0.4.2
[28] plotly_4.10.4     cli_3.6.3.9000    withr_3.0.2
[31] magrittr_2.0.3    crosstalk_1.2.1   digest_0.6.37
[34] grid_4.4.1        rstudioapi_0.17.1 lifecycle_1.0.4
[37] vctrs_0.6.5       evaluate_1.0.3    glue_1.8.0
[40] data.table_1.16.4 farver_2.1.2      pkgbuild_1.4.5
[43] colorspace_2.1-1  rmarkdown_2.29    purrr_1.0.2
[46] httr_1.4.7        tools_4.4.1       pkgconfig_2.0.3
[49] htmltools_0.5.8.1
```

Below is the result of `devtools::check()` on my computer :

```r
devtools::check(document = FALSE)

══ Building ═════════════════════════════════════════════════════════════════════════════════════════════════════════════════
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
── R CMD build ──────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file 'C:\Users\yohann\Documents\code\reviewropensci\mapmetadata/DESCRIPTION' (454ms)
─  preparing 'mapmetadata': (1.7s)
✔  creating vignettes (7.3s)
─  checking for LF line-endings in source and make files and shell scripts (891ms)
─  building 'mapmetadata_3.0.0.tar.gz'
   Avis dans utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/BAR_360_NationalCommunityChildHealthDatabase(NCCHD)_2024-12-19-14-11-55.html'
   Avis dans utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/CONSENSUS_MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-38-45.csv'
   Avis dans utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/L-MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv'
   Avis dans utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv'
   Avis dans utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv'
   Avis dans utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/MAPPING_LOG_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv'
   Avis dans utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/MAPPING_LOG_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv'
   Avis dans utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/MAPPING_PLOT_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.png'

══ Checking ═════════════════════════════════════════════════════════════════════════════════════════════════════════════════
Setting env vars:
• _R_CHECK_CRAN_INCOMING_REMOTE_               : FALSE
• _R_CHECK_CRAN_INCOMING_                      : FALSE
• _R_CHECK_FORCE_SUGGESTS_                     : FALSE
• _R_CHECK_PACKAGES_USED_IGNORE_UNUSED_IMPORTS_: FALSE
• NOT_CRAN                                     : true
── R CMD check ──────────────────────────────────────────────────────────────────────────────────────────────────────────────
─  using log directory 'C:/Users/yohann/Documents/code/reviewropensci/mapmetadata.Rcheck' (480ms)
─  using R version 4.4.1 (2024-06-14 ucrt)
─  using platform: x86_64-w64-mingw32
─  R was compiled by
       gcc.exe (GCC) 13.2.0
       GNU Fortran (GCC) 13.2.0
─  running under: Windows 11 x64 (build 22631)
─  using session charset: UTF-8
─  using options '--no-manual --as-cran' (858ms)
✔  checking for file 'mapmetadata/DESCRIPTION'
─  checking extension type ... Package
─  this is package 'mapmetadata' version '3.0.0'
─  package encoding: UTF-8
✔  checking package namespace information
✔  checking package dependencies (1.2s)
✔  checking if this is a source package ...
✔  checking if there is a namespace
✔  checking for executable files (926ms)
✔  checking for hidden files and directories ...
N  checking for portable file names
   Found the following non-portable file paths:
     mapmetadata/inst/outputs/BAR_360_NationalCommunityChildHealthDatabase(NCCHD)_2024-12-19-14-11-55.html
     mapmetadata/inst/outputs/CONSENSUS_MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-38-45.csv
     mapmetadata/inst/outputs/L-MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv
     mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv
     mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv
     mapmetadata/inst/outputs/MAPPING_LOG_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv
     mapmetadata/inst/outputs/MAPPING_LOG_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv
     mapmetadata/inst/outputs/MAPPING_PLOT_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.png

   Tarballs are only required to store paths of up to 100 bytes and cannot
   store those of more than 256 bytes, with restrictions including to 100
   bytes for the final component.
   See section 'Package structure' in the 'Writing R Extensions' manual.
✔  checking serialization versions
✔  checking whether package 'mapmetadata' can be installed (9.6s)
N  checking installed package size ...
       outputs   4.4Mb
✔  checking package directory (357ms)
N  checking for future file timestamps (4.2s)
   unable to verify current time
✔  checking 'build' directory
✔  checking DESCRIPTION meta-information (349ms)
✔  checking top-level files
✔  checking for left-over files ...
✔  checking index information ...
✔  checking package subdirectories (1s)
✔  checking code files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded (994ms)
✔  checking whether the package can be loaded with stated dependencies (706ms)
✔  checking whether the package can be unloaded cleanly (780ms)
✔  checking whether the namespace can be loaded with stated dependencies (788ms)
✔  checking whether the namespace can be unloaded cleanly (1.1s)
✔  checking loading without being on the library search path (1.2s)
✔  checking dependencies in R code (2s)
✔  checking S3 generic/method consistency (1.1s)
✔  checking replacement functions (892ms)
✔  checking foreign function calls (790ms)
N  checking R code for possible problems (6.2s)
   empty_count: no visible binding for global variable 'Section'
   empty_count: no visible binding for global variable
     'Column.description'
   empty_count: no visible binding for global variable 'total_variables'
   empty_count: no visible binding for global variable 'no_count'
   empty_count: no visible binding for global variable 'Empty'
   end_plot: no visible binding for global variable 'domain_code'
   map_compare: no visible binding for global variable 'Section'
   map_compare: no visible binding for global variable 'data_element'
   map_compare: no visible global function definition for 'contains'
   metadata_map: no visible binding for global variable 'Section'
   metadata_map: no visible binding for global variable 'note'
   output_copy: no visible binding for global variable 'data_element'
     Column.description Empty Section contains data_element domain_code
     no_count note total_variables
✔  checking Rd files (571ms)
✔  checking Rd metadata ...
N  checking Rd line widths ...
   Rd file 'map_compare.Rd':
     \examples lines wider than 100 characters:
        demo_metadata_file <- system.file("inputs", "360_NationalCommunityChildHealthDatabase(NCCHD)_Structural_Metadata.csv", package = "mapme ... [TRUNCATED]

   Rd file 'map_convert.Rd':
     \examples lines wider than 100 characters:
        demo_output_csv <- "MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv"

   These lines will be truncated in the PDF manual.
✔  checking Rd cross-references ...
✔  checking for missing documentation entries (925ms)
✔  checking for code/documentation mismatches (2.4s)
✔  checking Rd \usage sections (1.2s)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
✔  checking contents of 'data' directory
✔  checking data for non-ASCII characters ...
✔  checking LazyData
✔  checking data for ASCII and uncompressed saves ...
✔  checking installed files from 'inst/doc' ...
✔  checking files in 'vignettes' ...
✔  checking examples (1.5s)
✔  checking for unstated dependencies in 'tests' ...
─  checking tests ...
✔  Running 'testthat.R' (4.7s)
✔  checking for unstated dependencies in vignettes (563ms)
✔  checking package vignettes ...
✔  checking re-building of vignette outputs (4.9s)
✔  checking for non-standard things in the check directory ...
✔  checking for detritus in the temp directory

   See
     'C:/Users/yohann/Documents/code/reviewropensci/mapmetadata.Rcheck/00check.log'

── R CMD check results ─────────────────────────────────────────────────────────────────────────────── mapmetadata 3.0.0 ────
Duration: 1m 0.1s

❯ checking for portable file names ... NOTE
  Found the following non-portable file paths:
    mapmetadata/inst/outputs/BAR_360_NationalCommunityChildHealthDatabase(NCCHD)_2024-12-19-14-11-55.html
    mapmetadata/inst/outputs/CONSENSUS_MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-38-45.csv
    mapmetadata/inst/outputs/L-MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv
    mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv
    mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv
    mapmetadata/inst/outputs/MAPPING_LOG_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv
    mapmetadata/inst/outputs/MAPPING_LOG_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv
    mapmetadata/inst/outputs/MAPPING_PLOT_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.png

  Tarballs are only required to store paths of up to 100 bytes and cannot
  store those of more than 256 bytes, with restrictions including to 100
  bytes for the final component.
  See section 'Package structure' in the 'Writing R Extensions' manual.

❯ checking installed package size ... NOTE
      outputs   4.4Mb

❯ checking for future file timestamps ... NOTE
  unable to verify current time

❯ checking R code for possible problems ... NOTE
  empty_count: no visible binding for global variable 'Section'
  empty_count: no visible binding for global variable
    'Column.description'
[…]
