### 2025-01-17 #674 @RayStick
Thank you so much for your comments @ymansiaux !  I won’t work on them today (or until the other review comes in) but I have a quick question -

> To align with the naming of the other exported functions, map_compare() and map_convert(), I would suggest renaming metadata_map() to map_metadata(). While I understand you may have concerns about potential confusion with the package name, I believe this change would harmonize the naming scheme, enhance clarity, and facilitate autocomplete for users.

I struggled with function naming and changed them a lot 😅
My logic here was following this guidance: https://devguide.ropensci.org/pkg_building.html#function-and-argument-naming
- `metadata_map` because 'metadata' can be thought of the data object and 'map' the verb
- `map_convert` because now the 'map' (well 'mappings') are the object, and the verb is 'convert'
- `map_compare` now the 'map' (well 'mappings') are the object, an the verb is 'compare'

Does that make sense? I am open to changing them! On reflection, I am using map to be an object and a verb so that is confusing.
