### 2025-01-17 #674 @ymansiaux
Yes that perfectly makes sense :-) Considering the object_verb() naming scheme I think you should keep the current function names.

Have a good day !
