### 2025-01-17 #674 @maelle
As an alternative to `@noRd` you might enjoy https://github.com/moodymudskipper/devtag @RayStick (not in the dev guide yet), it generates Rd pages for internal functions, that developers can read with `?fun`, but Rbuildignores them so the users don't see them.
