+++ R/metadata_map.R
@@ -121,7 +121,7 @@ metadata_map <- function(
   ## Check if look_up_file and domain_file are compatible
   mistmatch <- setdiff(data$lookup$domain_code, df_plots$code$code)
   if (length(mistmatch) > 0) {
-    cli_alert_danger("The look_up_file and domain_file are not compatabile. These look up codes are not listed in the domain codes:")
+    cli_alert_danger("The look_up_file and domain_file are not compatible. These look up codes are not listed in the domain codes:")
     cat("\n")
     print(mistmatch)
     stop()
