### 2025-01-28 #674 @Lextuga007
Just to start by saying thank you for all the work you've put into this package 😃 Like @ymansiaux I learned a lot from the review and I hope my points are also helpful.

Some changes are required, particularly for README and vignette documentation.

---

Comprehensive tests for functions (both exported and internal) with every function having a test.
Good package name (checked with `pak`) which is descriptive.
Package has been built to interact as much as possible with the user and following that approach many of the subsequent comments are improvements from the documentation.
Really clear which functions are internal through the title.
Documentation fits the ROpenSci `principle of multiple points of entry` with README, vignettes and each function (including internal) documented.

## ROpenSci packaging guidelines

[Console messages](https://devguide.ropensci.org/pkg_building.html#console-messages) asks to avoid the use of `cat` and `print` as these are not easy to suppress.

Avoiding verbosity or suppressing messages is hard within this package as the interactions from the user require the messages.

## Accessibility

Thank you for the alt text on images!

To support further accessibility I recommend being more descriptive in the text where a link is highlighted in text like for the word [`here`](https://github.com/aim-rsf/mapmetadata/blob/13fb56e86239dfec3b8a8cc87d3c2ab68f2b4e5c/README.md?plain=1#L16) particularly as this is something that helps with [accessibility](https://www.wcag.com/blog/writing-meaningful-link-text/).

Also appears in the [vignette](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/vignettes/mapmetadata.Rmd#L34) which shows an image but only says `download here:`.

It took me a while to navigate the website as I wasn't familiar with it so I wonder if the image would benefit from a direct hyperlink to the same page?

And the non-descriptive `here` in [`see here for outputs generated from the demo run`](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/vignettes/mapmetadata.Rmd#L192).

## Technical language

It might be useful to expand on the point about changing the save location by [adjusting the `output_dir` argument](https://github.com/aim-rsf/mapmetadata/blob/13fb56e86239dfec3b8a8cc87d3c2ab68f2b4e5c/README.md?plain=1#L68) because `argument` is a technical term that might not be familiar to researchers - perhaps using an example to expand on this would help?

## Expectations of functionality

Only the Return key works to continue where the instruction says to [`finish here, or press any other key to continue with mapping variables`](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/README.md?plain=1#L65).
Other keys don't appear to "finish" and also, to be very clear, need to be executed in the Console.

The demo requires initials to be entered but it isn't explained why it is required.
It gets recorded in the logs but this might not be apparent to an initial user of the package.

Demo mode is said to process only the first 20 variables `from the selected table` and it's not clear what the 20 variables refer to particularly when directed to [selecting data elements](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/vignettes/mapmetadata.Rmd#L50).
Are `data elements` the same as `variables` in this example? I got a bit confused with this so couldn't complete the instructions.

The parameters used in the [`metadata_map()`](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/README.md?plain=1#L82) function example don't appear to work.
The code is `metadata_map(csv_file = new_csv_file, domain_file = demo_domains_file)` but there isn't a parameter called `csv_file` - should this be `metadata_file`?
I first of all tried to use the [`14_Mortality (Death registration)_Metadata.csv`](https://healthdatagateway.org/en/dataset/14) data from the website but got the error because the use of structural metadata is only referred to in the making of `metadata`.
I found this link via the arguments help for `metadata_file` which is excellent documentation, however, will be missed I think so would be really great to be clear in the README file as well.
An enhancement may be to have a message coded if a file is missing `structural` in the file name that the error is because it's the wrong file?

[`get("look_up")`](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/vignettes/mapmetadata.Rmd#L118) should perhaps come earlier in the vignette because it's not possible to use this function when in the process of selecting variables from using [`metadata_map()`](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/vignettes/mapmetadata.Rmd#L50) that appears just before.

For the prompt `Enter one or more numbers separated by spaces and then ENTER, or 0 to cancel` I tried just using the Enter/Return key to finish which gave an output.
Pressing Enter/Return key though isn't a listed option and I just guessed it would work so possibly better to be explicit.

I recommend linking to the function reference for [`map_compare()` in tutorial](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/vignettes/mapmetadata.Rmd#L209).

I recommend being explicit in the argument help for metadata_file as [`?metadata`](https://github.com/Lextuga007/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/R/metadata_map.R#L20) won't return any files if the package hasn't been loaded: `?mapmetadata::metadata`. Although this is functionality in RStudio it could be missed by someone new to R and RStudio and their experience of not finding help might be incorrectly associated with the package.

The reference to [`data/metadata.rda`](https://github.com/Lextuga007/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/R/metadata_map.R#L19) might not make sense as packages are often just used through the `library()` so perhaps better to use refer to the code `mapmetadata::metadata`?

For `Using a custom domain list input` where domain file inputs are detailed as automatically appending so [`do not include these in your domain list`](https://github.com/Lextuga007/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/README.md?plain=1#L89) is this because they get replicated? It would be helpful to be explicit that this wouldn't be likely to break code but just duplicate.

For `Using a custom lookup table input (advanced)` I couldn't work out the corresponding example csv that the instruction [`ensure that all domain codes in the lookup file are also included in your *domain file*`](https://github.com/Lextuga007/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/README.md?plain=1#L93) without opening all the csvs and looking for the column name.
It might be good to expand on this with an example.

Function documentation sometimes includes references to other functions in the package, for example [`valid_comparison()`](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/R/compare.R#L3) refers to it being used in the `the map_compare.R function`. Cross links are mentioned in the [ROpenSci package guidelines](https://devguide.ropensci.org/pkg_building.html#roxygen-2-use) and if you use RStudio I recommend using my colleague, Matt Dray's, [snorkel](https://github.com/matt-dray/snorkel) package to add these in as you can highlight the text and use the RStudio add in.

## Spelling

Minor spelling have been submitted for [correction on a PR](https://github.com/aim-rsf/mapmetadata/pull/183) but where functions have spelling mistakes:

`concensus` is a function name and will require updates to corresponding files: `concensus_on_mismatch()`, `test-concensus_on_mismatch.R`.

It is also in the message from the [`map_compare()` function](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/R/compare.R#L57) and is in the documentation for `map_compare()`.

Object in function `metadata_map()` (which won't have any effect to the user) `mistmatch`.

## Package build

Notes on `no visible binding for global variable` can be managed using the utils package and a specific R file (`globals.R`). More details on more [sophisticated treatment](https://r-pkgs.org/package-within.html#echo-a-working-package) can be found in the R packages book. I tend to get these issues particularly when referring to SQL columns in R code.

`portable file names... NOTE` is appearing because the file names are over 100 characters long.
These files are referred to in the `map_convert()` and `map_compare()` documentation example so it might be ok to rename this longer name by the parts for generalising? For example:

```
#' @examples
#' # Locate file path and file name for the example files in the package
#' demo_output_dir <- system.file("outputs", package = "mapmetadata")
#' demo_output_csv <- "MAPPING_360_example_datetime.csv"
#'
#' # Run the function
#' map_convert(output_csv = demo_output_csv, output_dir = demo_output_dir)
```

## Formatting consistency

I noticed that there are line breaks put into function `metadata_map.R` so almost all the text is within an 80 character limit, but this isn't the case for other scripts.
It's quite nice to have a line break set because it makes it easier to read across the screen and as it's been used in one script I wondered if it could be applied to all scripts?
I use `lintr` to highlight the lines that are over 80 characters.

@ymansiaux shared the devtools::check output which is useful as I don't get all the same notes (I'm missing the line widths note) so for comparison this is mine:

```
══ Documenting ══════════════════════════════════════════════════
ℹ Updating mapmetadata documentation
ℹ Loading mapmetadata

══ Building ═════════════════════════════════════════════════════
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
── R CMD build ──────────────────────────────────────────────────
✔  checking for file 'C:\Users\zoe.turner\OneDrive - Midlands and Lancashire CSU\Documents\4. GitHub\mapmetadata/DESCRIPTION' (403ms)
─  preparing 'mapmetadata': (536ms)
✔  creating vignettes (9.8s)
─  checking for LF line-endings in source and make files and shell scripts (1.3s)
─  building 'mapmetadata_3.0.0.tar.gz'
   Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/BAR_360_NationalCommunityChildHealthDatabase(NCCHD)_2024-12-19-14-11-55.html'
   Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/CONSENSUS_MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-38-45.csv'
   Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/L-MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv'
   Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv'
   Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv'
   Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/MAPPING_LOG_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv'
   Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/MAPPING_LOG_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv'
   Warning in utils::tar(filepath, pkgname, compression = compression, compression_level = 9L,  :
     storing paths of more than 100 bytes is not portable:
     'mapmetadata/inst/outputs/MAPPING_PLOT_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.png'

══ Checking ═════════════════════════════════════════════════════
Setting env vars:
• _R_CHECK_CRAN_INCOMING_REMOTE_               : FALSE
• _R_CHECK_CRAN_INCOMING_                      : FALSE
• _R_CHECK_FORCE_SUGGESTS_                     : FALSE
• _R_CHECK_PACKAGES_USED_IGNORE_UNUSED_IMPORTS_: FALSE
• NOT_CRAN                                     : true
── R CMD check ──────────────────────────────────────────────────
─  using log directory 'C:/Users/zoe.turner/AppData/Local/Temp/Rtmp4SNDF6/file83c42e8f7965/mapmetadata.Rcheck' (536ms)
─  using R version 4.4.1 (2024-06-14 ucrt)
─  using platform: x86_64-w64-mingw32
─  R was compiled by
       gcc.exe (GCC) 13.2.0
       GNU Fortran (GCC) 13.2.0
─  running under: Windows 10 x64 (build 19045)
─  using session charset: UTF-8
─  using options '--no-manual --as-cran'
✔  checking for file 'mapmetadata/DESCRIPTION' ...
─  checking extension type ... Package
─  this is package 'mapmetadata' version '3.0.0'
─  package encoding: UTF-8
✔  checking package namespace information
✔  checking package dependencies (1.6s)
✔  checking if this is a source package ...
✔  checking if there is a namespace
✔  checking for executable files (1.4s)
✔  checking for hidden files and directories ...
N  checking for portable file names
   Found the following non-portable file paths:
     mapmetadata/inst/outputs/BAR_360_NationalCommunityChildHealthDatabase(NCCHD)_2024-12-19-14-11-55.html
     mapmetadata/inst/outputs/CONSENSUS_MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-38-45.csv
     mapmetadata/inst/outputs/L-MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv
     mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv
     mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv
     mapmetadata/inst/outputs/MAPPING_LOG_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv
     mapmetadata/inst/outputs/MAPPING_LOG_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv
     mapmetadata/inst/outputs/MAPPING_PLOT_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.png

   Tarballs are only required to store paths of up to 100 bytes and cannot
   store those of more than 256 bytes, with restrictions including to 100
   bytes for the final component.
   See section 'Package structure' in the 'Writing R Extensions' manual.
✔  checking serialization versions
✔  checking whether package 'mapmetadata' can be installed (6.2s)
N  checking installed package size (443ms)
       outputs   4.4Mb
✔  checking package directory (378ms)
✔  checking for future file timestamps ...
✔  checking 'build' directory
✔  checking DESCRIPTION meta-information (394ms)
✔  checking top-level files
✔  checking for left-over files ...
✔  checking index information (384ms)
✔  checking package subdirectories (1.3s)
✔  checking code files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded (1.8s)
✔  checking whether the package can be loaded with stated dependencies (962ms)
✔  checking whether the package can be unloaded cleanly (1.3s)
✔  checking whether the namespace can be loaded with stated dependencies (997ms)
✔  checking whether the namespace can be unloaded cleanly (1.2s)
✔  checking dependencies in R code (2.3s)
✔  checking S3 generic/method consistency (1.2s)
✔  checking replacement functions (992ms)
✔  checking foreign function calls (1.1s)
N  checking R code for possible problems (8.7s)
   empty_count: no visible binding for global variable 'Section'
   empty_count: no visible binding for global variable
     'Column.description'
   empty_count: no visible binding for global variable 'total_variables'
   empty_count: no visible binding for global variable 'no_count'
   empty_count: no visible binding for global variable 'Empty'
   end_plot: no visible binding for global variable 'domain_code'
   map_compare: no visible binding for global variable 'Section'
   map_compare: no visible binding for global variable 'data_element'
   map_compare: no visible global function definition for 'contains'
   metadata_map: no visible binding for global variable 'Section'
   metadata_map: no visible binding for global variable 'note'
   output_copy: no visible binding for global variable 'data_element'
     Column.description Empty Section contains data_element domain_code
     no_count note total_variables
✔  checking Rd files (533ms)
✔  checking Rd metadata ...
N  checking Rd line widths (342ms)
   Rd file 'map_compare.Rd':
     \examples lines wider than 100 characters:
        demo_metadata_file <- system.file("inputs", "360_NationalCommunityChildHealthDatabase(NCCHD)_Structural_Metadata.csv", package = "mapme ... [TRUNCATED]

   Rd file 'map_convert.Rd':
     \examples lines wider than 100 characters:
        demo_output_csv <- "MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv"

   These lines will be truncated in the PDF manual.
✔  checking Rd cross-references ...
✔  checking for missing documentation entries (1.1s)
✔  checking for code/documentation mismatches (3.7s)
✔  checking Rd \usage sections (1.8s)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
✔  checking contents of 'data' directory ...
✔  checking data for non-ASCII characters (445ms)
✔  checking LazyData
✔  checking data for ASCII and uncompressed saves ...
✔  checking installed files from 'inst/doc' (410ms)
✔  checking files in 'vignettes' (435ms)
✔  checking examples (1.9s)
✔  checking for unstated dependencies in 'tests' ...
─  checking tests ...
✔  Running 'testthat.R' (4.7s)
✔  checking for unstated dependencies in vignettes (666ms)
✔  checking package vignettes ...
✔  checking re-building of vignette outputs (5.6s)
✔  checking for non-standard things in the check directory
✔  checking for detritus in the temp directory

   See
     'C:/Users/zoe.turner/AppData/Local/Temp/Rtmp4SNDF6/file83c42e8f7965/mapmetadata.Rcheck/00check.log'

── R CMD check results ───────────────────────────────────────────────────────────────────────────────────────────────────── mapmetadata 3.0.0 ────
Duration: 1m 0.7s

❯ checking for portable file names ... NOTE
  Found the following non-portable file paths:
    mapmetadata/inst/outputs/BAR_360_NationalCommunityChildHealthDatabase(NCCHD)_2024-12-19-14-11-55.html
    mapmetadata/inst/outputs/CONSENSUS_MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-38-45.csv
    mapmetadata/inst/outputs/L-MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv
    mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-11-55.csv
    mapmetadata/inst/outputs/MAPPING_360_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-12-19-14-17-45.csv
[…]
