### 2025-01-29 #674 @maelle
Thanks a lot @Lextuga007 for your thorough review! 🚀

Regarding verbosity, we've published a blog post a while ago (with the goal of updating the dev guide later), that might be relevant: https://ropensci.org/blog/2024/02/06/verbosity-control-packages/

@RayStick Both reviews are now in. 😸
