### 2025-01-30 9992ba0
Update CITATION.cff
