### 2025-02-02 #674 @maurolepore
Dear @RayStick this is to mark the start of my EiC rotation. I'm reviewing all open issues and leaving a short note to myself of what I see.

I see both reviews are in and I assume you're working on incorporating the feedback. Everything seems on track. I'll step back. Thanks!
