### 2025-02-03 aim-rsf/mapmetadata#185 @RayStick
Improve accessibility of links & clarify technical language & correct spelling (Reviewer 2, points 2 & 3 & 5)
**[2] Accessibility**

Thank you for the alt text on images!

To support further accessibility I recommend being more descriptive in the text where a link is highlighted in text like for the word [`here`](https://github.com/aim-rsf/mapmetadata/blob/13fb56e86239dfec3b8a8cc87d3c2ab68f2b4e5c/README.md?plain=1#L16) particularly as this is something that helps with [accessibility](https://www.wcag.com/blog/writing-meaningful-link-text/).

Also appears in the [vignette](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/vignettes/mapmetadata.Rmd#L34) which shows an image but only says `download here:`.

It took me a while to navigate the website as I wasn't familiar with it so I wonder if the image would benefit from a direct hyperlink to the same page?

And the non-descriptive `here` in [`see here for outputs generated from the demo run`](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/vignettes/mapmetadata.Rmd#L192).

**[3] Technical language**

It might be useful to expand on the point about changing the save location by [adjusting the `output_dir` argument](https://github.com/aim-rsf/mapmetadata/blob/13fb56e86239dfec3b8a8cc87d3c2ab68f2b4e5c/README.md?plain=1#L68) because `argument` is a technical term that might not be familiar to researchers - perhaps using an example to expand on this would help?

**[5] Spelling**

Minor spelling have been submitted for https://github.com/aim-rsf/mapmetadata/pull/183 but where functions have spelling mistakes:

concensus is a function name and will require updates to corresponding files: concensus_on_mismatch(), test-concensus_on_mismatch.R.

It is also in the message from the [map_compare() function](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/R/compare.R#L57) and is in the documentation for map_compare().

Object in function metadata_map() (which won't have any effect to the user) mistmatch.
