### 2025-02-03 fbd4586
clarify technical language
