### 2025-02-03 aim-rsf/mapmetadata#184 @maelle
> is there also an editor contribution tag for @maelle

No, thank you for asking. :pray: In the dev guide we state "Please do not list editors as contributors. Your participation in and contribution to rOpenSci is thanks enough!" https://devguide.ropensci.org/pkg_building.html#authorship (I'm actually realizing you listed me as contributor in some other metadata!)
