### 2025-02-03 aim-rsf/mapmetadata#184 @RayStick
> No, thank you for asking. 🙏 In the dev guide we state "Please do not list editors as contributors. Your participation in and contribution to rOpenSci is thanks enough!" https://devguide.ropensci.org/pkg_building.html#authorship (I'm actually realizing you listed me as contributor in some other metadata!)
Ah, missed that guide note! Well I am happy to keep you as a contributor (if you are okay with that) as I *think* you gave ideas - and contributions in this repo - above and beyond your editor role, but correct me if I am wrong. Thanks!
