### 2025-02-10 ac92b44
shorten file names and troubleshoot linting that broke things
