+++ man/metadata_map.Rd
@@ -15,7 +15,7 @@ metadata_map(
 }
 \arguments{
 \item{metadata_file}{This should be a csv download from HDRUK gateway
-(0_Dataset_Structural_Metadata.csv). Default is 'data/metadata.rda' - run
+(in the form of ID_Dataset_Metadata.csv). Default is 'data/metadata.rda': run
 '?metadata' to see how it was created.}
 
 \item{domain_file}{This should be a csv file created by the user, with each

[…]
