### 2025-02-10 104f437
shorten and fix linting lines
