### 2025-02-10 2eee4ca
remove large html file from package
