### 2025-02-10 7958ad9
add global variables
