+++ R/globals.R
@@ -1,0 +1,9 @@
+utils::globalVariables(c("Column.description",
+                         "Empty",
+                         "Section",
+                         "contains",
+                         "data_element",
+                         "domain_code",
+                         "no_count",
+                         "note",
+                         "total_variables"))
