### 2025-02-10 db64711
clarify arguments and improve example with tempdir()
