### 2025-02-10 f1cc462
use interactive mode and tempdir()
