### 2025-02-11 #674 @maelle
Yes it is ok @RayStick, no worries.

Note that once you post your response, you can "submit" it with a bot command so the label of the issue will change: https://devguide.ropensci.org/bot_cheatsheet.html#submit-response-to-reviewers
