### 2025-02-12 52d9c5f
correct the way metadata_map calls map_convert
