### 2025-02-12 4e2ff2a
remove unnecessary cat and print
