+++ R/map_convert.R
@@ -44,7 +44,7 @@ map_convert <- function(csv_to_convert,
   }
 
   # Save output_long
-  output_long_fname <- paste0(output_dir, "L-", csv_to_convert)
+  output_long_fname <- paste0(output_dir, "/L-", csv_to_convert)
   write.csv(output_long, output_long_fname, row.names = FALSE)
   cat("\n")
   cli_alert_success(paste("Long format categorisations saved to:\n",
