### 2025-02-12 58dd20f
remove unnecessary cat and print
