+++ R/metadata_map.R
@@ -79,8 +79,8 @@ metadata_map <- function(
   n_tables <- length(levels(dataset$Section))
 
   ## Print info about dataset to user
-  cli_alert_info("Processing dataset: {dataset_name}")
-  cli_alert_info("There are {n_tables} tables in this dataset")
+  cli_alert_info(paste("Processing dataset '{dataset_name}' containing",
+                       "{n_tables} tables\n\n"))
 
   # SECTION 2 - CREATE SUMMARY BAR PLOT FOR DATASET ----
 
@@ -103,19 +103,13 @@ metadata_map <- function(
   setwd(original_wd) # saveWidget has a bug with paths & saving
 
   ## Display outputs to the user
-  cat("\n")
   browseURL(file.path(output_dir, bar_fname))
-  cli_alert_info(paste("A bar plot should have opened in your browser.",
-                       "It has also been saved to your project directory",
-                       "(alongside a csv).Use this bar plot, and the",
-                       "information on the HDRUK Gateway, to guide your",
-                       "mapping approach."))
+  cli_alert_info(paste("A bar plot should have opened in your browser",
+                       "(it has also been saved to your project directory).\n",
+                       "Use this bar plot, and the information on the HDRUK",
+                       "Gateway, to guide your mapping approach.\n\n"))
 
   # SECTION 3 - MAPPING VARIABLES TO CONCEPTS (DOMAINS) FOR EACH TABLE ----
-
-  cat("\n")
-  readline(paste("Press 'Esc' key to finish here, or press any other key to",
-                 "continue with mapping variables"))
 
   ## Read in prepared output data frames
   log_output_df <- get("log_output_df")
@@ -128,14 +122,10 @@ metadata_map <- function(
   mismatch <- setdiff(data$lookup$domain_code, df_plots$code$code)
   if (length(mismatch) > 0) {
     cli_alert_danger("The look_up_file and domain_file are not compatible.
-                     These look up codes are not listed in the domain codes:")
-    cat("\n")
+                     These look up codes are not listed in the domain codes:\n")
     print(mismatch)
     stop()
   }
-
-  ## Get user initials for the log file
-  user_initials <- readline(prompt = "Enter your initials: ")
 
   ## CHOOSE TABLE TO PROCESS
 
@@ -143,10 +133,8 @@ metadata_map <- function(
                          levels(dataset$Section),
                          title = "Enter the table number you want to process:")
   table_name <- levels(dataset$Section)[chosen_table_n]
-  cat("\n")
-  cli_alert_info("Processing Table {chosen_table_n} of {n_tables}
-                 ({table_name})")
-  cat("\n")
+  cli_alert_info(paste("Processing Table {chosen_table_n} of {n_tables}",
+                       "({table_name})\n\n"))
 
   #### Use 'output_copy.R' to copy from previous output(s) if they exist
   if (table_copy == TRUE) {
@@ -157,10 +145,7 @@ metadata_map <- function(
     df_prev_exist <- FALSE
   }
 
-  table_note <- readline(paste(
-    "Optional free text note about this table",
-    "(or press 'Enter'): "
-  ))
+  table_note <- readline(paste("Optional note about this table: "))
 
   ####  Extract table from metadata
   table_df <- dataset %>%
@@ -192,9 +177,7 @@ metadata_map <- function(
 
 
   #### Review auto categorized data elements
-  cat("\n")
-  cli_alert_info("These are the auto categorised data elements:")
-  cat("\n")
+  cli_alert_info("These are the auto categorised data elements:\n\n")
   output_auto <- subset(output_df, note == "AUTO CATEGORISED")
[…]
