### 2025-02-12 687d297
check cli package and add quiet option
