+++ R/user_interactions.R
@@ -94,6 +94,8 @@ user_categorisation <- function(data_element, data_desc, data_type,
 #' @param lookup The lookup table to enable auto categorisations
 #' @param df_plots Output from ref_plot, to indicate mac domain code allowed
 #' @param output_df Empty output dataframe, to fill
+#' @param quiet Default is FALSE. Change to TRUE to quiet the cli_alert_info
+#' and cli_alert_success messages.
 #' @return An output dataframe containing info about the table, data elements
 #' and categorisations
 #' @keywords internal
@@ -101,10 +103,13 @@ user_categorisation <- function(data_element, data_desc, data_type,
 #' @importFrom cli cli_alert_info
 
 user_categorisation_loop <- function(start_v, end_v, table_df, df_prev_exist,
-                                     df_prev, lookup, df_plots, output_df) {
+                                     df_prev, lookup, df_plots, output_df,
+                                     quiet = FALSE) {
   for (data_v in start_v:end_v) {
-    cli_alert_info(paste(length(data_v:end_v), "left to process",
-                         "(Data element {data_v} of {nrow(table_df)})"))
+    if (!quiet) {
+      cli_alert_info(paste(length(data_v:end_v), "left to process",
+                           "(Data element {data_v} of {nrow(table_df)})"))
+    }
     this_data_element <- table_df$Column.name[data_v]
     this_data_element_n <- paste(
       as.character(data_v), "of",
