+++ R/user_interactions.R
@@ -107,8 +107,8 @@ user_categorisation_loop <- function(start_v, end_v, table_df, df_prev_exist,
                                      quiet = FALSE) {
   for (data_v in start_v:end_v) {
     if (!quiet) {
-      cli_alert_info(paste(length(data_v:end_v), "left to process",
-                           "(Data element {data_v} of {nrow(table_df)})"))
+      cli_alert_info(paste0("Data element {data_v} of {nrow(table_df)} (",
+                           length(data_v:end_v), " left to process)"))
     }
     this_data_element <- table_df$Column.name[data_v]
     this_data_element_n <- paste(
