### 2025-02-12 f0c948c
update docs to match code changes
