+++ vignettes/mapmetadata.Rmd
@@ -31,7 +31,7 @@ It is important to note that this plot is only summarising *variable* level meta
 
 The bar plot can help you understand the scope of the dataset, but reference the HDRUK Gateway page for the fuller context. For instance, table descriptions are not included in these structural metadata files but they are included on the gateway:
 
-For dataset NCCHD, used in the demo, the structural metadata was downloaded from https://healthdatagateway.org/en/dataset/360. Below, we include a screenshot of this page; a user must click on the 'Download Data' then 'Structural Metadata' in order to download the file. Please note, we are working on API integration with this package to streamline this for users.
+For dataset NCCHD, used in the demo, the structural metadata was downloaded from <https://healthdatagateway.org/en/dataset/360>. Below, we include a screenshot of this page; a user must click on the 'Download Data' then 'Structural Metadata' in order to download the file. Please note, we are working on API integration with this package to streamline this for users.
 
 <img src="https://raw.githubusercontent.com/aim-rsf/mapmetadata/main/inst/outputs/NCCHD_gateway_download.png" alt="screenshot of HRDRUK gateway showing location to download structural metadata"/>
 
@@ -43,14 +43,15 @@ It is recommended to view the descriptions of each table to give you more contex
 
 Use the bar plot and the HDRUK Gateway to guide your mapping choices. The main functionality of this package is to aid a researcher in mapping variables from health datasets onto their research domains (concepts/latent variables).
 
-You will see this in the R console:
-
-`Press 'Esc' key to finish here, or press any other key to continue with mapping variables`
-
-Pressing any key will move you to the mapping phase. Demo mode (run by `metadata_map()`) only processes the first 20 data elements (variables) in the selected table.
+Run the demo mode by typing `metadata_map()`. You you will see this in the R console:
 
 ```         
-Enter your initials: rs
+ℹ Running demo mode using package data files
+ℹ Using the default look-up table in data/look-up.rda
+ℹ Processing dataset '360_NCCHD' containing 13 tables
+
+ℹ A bar plot should have opened in your browser (also saved to your project directory).
+Use this bar plot, and the information on the HDRUK Gateway, to guide your mapping approach.
 
 Enter the table number you want to process: 
 
@@ -69,11 +70,17 @@ Enter the table number you want to process:
  13: SIG_COND                        
 
 Selection: 4
+```
 
+Enter 4 to follow this demo. Demo mode only processes the first 20 data elements (variables) in the selected table.
+
+```         
 ℹ Processing Table 4 of 13 (CHILD)
 
-Optional free text note about this table (or press enter to continue): tutorial testing
+Optional note about this table: tutorial testing
 ```
+
+You can include a note about this table (e.g. project name, who is processing), or press enter to leave it blank.
 
 You will be asked to label data elements with one (or more) of the numbers shown in the Plots tab [0-7]. Here we have very simple domains [4-7] for the demo run. For a research study, your domains are likely to be much more specific e.g. 'Prenatal, antenatal, neonatal and birth' or 'Health behaviours and diet'. The 4 default domains are always included [0-3], appended on to any domain list given by the user.
 
@@ -82,27 +89,19 @@ You will be asked to label data elements with one (or more) of the numbers shown
 If it skips over a data element it means it was auto-categorised or copied from a previous table already processed (more on that later).
 
 ```         
-ℹ 20 left to process in this session
-✔ Processing data element 1 of 35
[…]
