+++ R/compare.R
@@ -57,11 +57,11 @@ consensus_on_mismatch <- function(ses_join, table_df, datavar,
     cat("\n\n")
     cli_alert_warning("Mismatch found, provide concensus decision below.")
     cli_alert_warning(paste("\nDOMAIN CODE (note) for session 1 --> ",
-                      ses_join$domain_code_ses1[datavar],
-                      "(", ses_join$note_ses1[datavar], ")"))
+                            ses_join$domain_code_ses1[datavar],
+                            "(", ses_join$note_ses1[datavar], ")"))
     cli_alert_warning(paste("\nDOMAIN CODE (note) for session 2 --> ",
-                      ses_join$domain_code_ses2[datavar],
-                      "(", ses_join$note_ses2[datavar], ")\n"))
+                            ses_join$domain_code_ses2[datavar],
+                            "(", ses_join$note_ses2[datavar], ")\n"))
     decision_output <- user_categorisation(table_df$Column.name[datavar],
                                            table_df$Column.description[datavar],
                                            table_df$Data.type[datavar],
