+++ man/ref_plot.Rd
@@ -21,3 +21,4 @@ This reference table is based on the user inputted domains and the default
 domains provided by this package.  \cr \cr
 }
 \keyword{internal}
+\keyword{internal}
