### 2025-02-14 7f4aa94
simplify demo run and check for previous bar plot
