### 2025-02-14 51e1190
re-arrange and simplify
