+++ tests/testthat/test-consensus_on_mismatch.R
@@ -1,8 +1,8 @@
 test_that("consensus_on_mismatch handles mismatch correctly", {
   # Mock the user_categorisation function
-  local_mocked_bindings(user_categorisation = function(data_element = NULL,
-                                                       data_desc = NULL,
-                                                       data_type = NULL,
+  local_mocked_bindings(user_categorisation = function(var = NULL,
+                                                       desc = NULL,
+                                                       type = NULL,
                                                        domain_code_max = NULL) {
     return(list(decision = "mock_decision", decision_note = "mock_note"))
   })
@@ -11,24 +11,24 @@ test_that("consensus_on_mismatch handles mismatch correctly", {
   ses_join <- data.frame(
     domain_code_ses1 = c("1", "1,2"),
     domain_code_ses2 = c("1", "2"),
-    data_element = c("Element1", "Element2"),
+    variable = c("Variable1", "Variable2"),
     note_ses1 = c("Note1", "Note2"),
     note_ses2 = c("Note3", "Note4"),
     stringsAsFactors = FALSE
   )
 
   table_df <- data.frame(
-    label = c("Element1", "Element2"),
+    label = c("Variable1", "Variable2"),
     description = c("Description1", "Description2"),
     type = c("Type1", "Type2"),
     stringsAsFactors = FALSE
   )
 
-  datavar <- 2
+  table_var <- 2
   domain_code_max <- 5
 
   # Call the function
-  result <- consensus_on_mismatch(ses_join, table_df, datavar, domain_code_max)
+  result <- consensus_on_mismatch(ses_join, table_df, table_var, domain_code_max)
 
   # Check the result
   expect_equal(result$domain_code_join, "mock_decision")
@@ -40,24 +40,24 @@ test_that("consensus_on_mismatch handles no mismatch correctly", {
   ses_join <- data.frame(
     domain_code_ses1 = c("2", "2,4"),
     domain_code_ses2 = c("2", "2,4"),
-    data_element = c("Element1", "Element2"),
+    variable = c("Variable1", "Variable2"),
     note_ses1 = c("Note1", "Note2"),
     note_ses2 = c("Note3", "Note4"),
     stringsAsFactors = FALSE
   )
 
   table_df <- data.frame(
-    label = c("Element1", "Element2"),
+    label = c("Variable1", "Variable2"),
     description = c("Description1", "Description2"),
     type = c("Type1", "Type2"),
     stringsAsFactors = FALSE
   )
 
-  datavar <- 2
+  table_var <- 2
   domain_code_max <- 5
 
   # Call the function
-  result <- consensus_on_mismatch(ses_join, table_df, datavar, domain_code_max)
+  result <- consensus_on_mismatch(ses_join, table_df, table_var, domain_code_max)
 
   # Check the result
   expect_equal(result$domain_code_join, "2,4")
