### 2025-02-14 9172429
input checks (R1-3 and R2-4d)
