+++ R/inputs_collect.R
@@ -5,21 +5,26 @@
 #' If some inputs are NULL, it loads the default inputs. If defaults not
 #' available, it prints error for the user.
 #' If inputs are not NULL, it loads the user-specified inputs.
-#' @param csv_file As defined in metadata_map
+#' @param metadata_file As defined in metadata_map
 #' @param domain_file As defined in metadata_map
 #' @param look_up_file As defined in metadata_map
 #' @param quiet Default is FALSE. Change to TRUE to quiet the cli_alert_info
 #' and cli_alert_success messages.
 #' @return A list of 6: all inputs needed for the metadata_map function to run.
-#' @importFrom cli cli_alert_info cli_alert_danger
+#' @importFrom cli cli_alert_info
 #' @importFrom utils read.csv
 #' @importFrom tools file_path_sans_ext
 #' @keywords internal
 #' @dev generate help files for unexported objects, for developers
 
-data_load <- function(csv_file, domain_file, look_up_file, quiet = FALSE) {
-  # Collect metadata and domains
-  if (is.null(csv_file) && is.null(domain_file)) {
+data_load <- function(metadata_file, domain_file, look_up_file, quiet = FALSE) {
+
+  ## Check that quiet is a boolean
+  if (!is.logical(quiet)) {
+    stop(paste("quiet should take the value of 'TRUE' or 'FALSE'"))
+  }
+
+  if (is.null(metadata_file) && is.null(domain_file)) {
     metadata <- get("metadata")
     metadata_desc <- "360_NCCHD"
     domains <- get("domain_list")
@@ -28,17 +33,48 @@ data_load <- function(csv_file, domain_file, look_up_file, quiet = FALSE) {
       cli_alert_info("Running demo mode using package data files")
     }
     demo_mode <- TRUE
-  } else if (is.null(csv_file) || is.null(domain_file)) {
-    cli_alert_danger("Provide both csv & domain file (or neither to run demo)")
-    stop()
-  } else { # read in user specified files
+  } else if (is.null(metadata_file) && !is.null(domain_file) ||
+               !is.null(metadata_file) && is.null(domain_file)) {
+    stop("Provide both csv & domain file (or neither to run demo)")
+  } else {
+    # read in user specified files
     demo_mode <- FALSE
-    metadata <- read.csv(csv_file)
-    metadata_base <- basename(csv_file)
+    metadata_base <- basename(metadata_file)
+
+    # Verify the metadata file name pattern and that it is a .csv file
+    if (!grepl("^[0-9]+_.*_Metadata\\.csv$", metadata_base) ||
+          tools::file_ext(metadata_file) != "csv") {
+      stop(paste("Metadata file name must be a .csv file in the format",
+                 "ID_Name_Metadata.csv where ID is an integer"))
+    } else {
+      if (!file.exists(metadata_file)) {
+        stop("Metadata filename is the correct format but it does not exist!")
+      }
+    }
+
+    # Check if metadata column names match what is expected
+    metadata <- read.csv(metadata_file)
+    column_names <- colnames(metadata)
+    expected_column_names <- c("Section", "Column.name", "Data.type",
+                               "Column.description", "Sensitive")
+    if (!all(column_names == expected_column_names)) {
+      stop("Metadata file does not have expected column names")
+    }
+
     metadata_base_0suffix <- sub("_Metadata.csv$", "", metadata_base)
     metadata_desc <- gsub(" ", "", metadata_base_0suffix)
-    domains <- read.csv(domain_file, header = FALSE)
-    domain_list_desc <- file_path_sans_ext(basename(domain_file))
+
+    # Check if the domain_file is a csv and has one column
+    if (file.exists(domain_file) && tools::file_ext(domain_file) == "csv") {
[…]
