+++ R/map_compare.R
@@ -55,27 +55,76 @@ map_compare <- function(session_dir,
                         quiet = FALSE) {
   timestamp_now_fname <- format(Sys.time(), "%Y-%m-%d-%H-%M-%S")
 
-  # DEFINE INPUTS ----
+  # VALIDATE & DEFINE INPUTS ----
 
-  csv_1a <- read.csv(file.path(session_dir,
-                               paste0("MAPPING_LOG_", session1_base, ".csv")))
-  csv_2a <- read.csv(file.path(session_dir,
-                               paste0("MAPPING_LOG_", session2_base, ".csv")))
-  csv_1b <- read.csv(file.path(session_dir,
-                               paste0("MAPPING_", session1_base, ".csv")))
-  csv_2b <- read.csv(file.path(session_dir,
-                               paste0("MAPPING_", session2_base, ".csv")))
+  ## Check that quiet is a boolean
+  if (!is.logical(quiet)) {
+    stop(paste("quiet should take the value of 'TRUE' or 'FALSE'"))
+  }
 
-  dataset <- read.csv(metadata_file)
-  domains <- read.csv(domain_file, header = FALSE)
+  ## Check if output_dir exists
+  if (!dir.exists(output_dir)) {
+    stop("The output_dir does not exist.")
+  }
 
-  ## Extract name of dataset
-  metadata_file_base <- basename(metadata_file)
-  metadata_file_base_0suffix <- sub("_Metadata.csv$", "",
-                                    metadata_file_base)
-  dataset_name <- gsub(" ", "", metadata_file_base_0suffix)
+  # Define csv file paths
+  csv_1a_path <- file.path(session_dir, paste0("MAPPING_LOG_", session1_base,
+                                               ".csv"))
+  csv_2a_path <- file.path(session_dir, paste0("MAPPING_LOG_", session2_base,
+                                               ".csv"))
+  csv_1b_path <- file.path(session_dir, paste0("MAPPING_", session1_base,
+                                               ".csv"))
+  csv_2b_path <- file.path(session_dir, paste0("MAPPING_", session2_base,
+                                               ".csv"))
 
-  # VALIDATION CHECKS ----
+  ## Check if csv files exist
+  if (!file.exists(csv_1a_path) || !file.exists(csv_2a_path)
+      || !file.exists(csv_1b_path) || !file.exists(csv_2b_path))  {
+    stop("Cannot locate all four input files.")
+  }
+
+  # Read csv files
+  csv_1a <- read.csv(csv_1a_path)
+  csv_2a <- read.csv(csv_2a_path)
+  csv_1b <- read.csv(csv_1b_path)
+  csv_2b <- read.csv(csv_2b_path)
+
+  # Verify the metadata file name pattern and that it is a .csv file
+  metadata_base <- basename(metadata_file)
+  if (!grepl("^[0-9]+_.*_Metadata\\.csv$", metadata_base) ||
+        tools::file_ext(metadata_file) != "csv") {
+    stop(paste("Metadata file name must be a .csv file in the format",
+               "ID_Name_Metadata.csv where ID is an integer"))
+  } else {
+    if (!file.exists(metadata_file)) {
+      stop("Metadata filename is the correct format but it does not exist!")
+    }
+  }
+
+  # Check if metadata column names match what is expected
+  metadata <- read.csv(metadata_file)
+  column_names <- colnames(metadata)
+  expected_column_names <- c("Section", "Column.name", "Data.type",
+                             "Column.description", "Sensitive")
+  if (!all(column_names == expected_column_names)) {
+    stop("Metadata file does not have expected column names")
+  }
+
[…]
