+++ vignettes/mapmetadata.Rmd
@@ -20,7 +20,7 @@ For installation, set-up and basic usage refer to the package [README.md](https:
 
 <img src="https://raw.githubusercontent.com/aim-rsf/mapmetadata/main/inst/outputs/BAR_360_NCCHD_2024-12-19-14-11-55.png" alt="example bar plot showing number of variables for each table alongside counts of whether variables have missing descriptions"/>
 
-This bar plot is produced automatically when you run the `[metadata_map](R/metadata_map.R)` function with the demo metadata file, which contains metadata about the **National Community Child Health Database (NCCHD).**
+This bar plot is produced automatically when you run the `[metadata_map](https://github.com/aim-rsf/mapmetadata/tree/main/R/metadata_map.R)` function with the demo metadata file, which contains metadata about the **National Community Child Health Database (NCCHD).**
 
 The bar plot shows us there are 13 tables in the dataset. The height of the bar indicates the number of variables in that table:
 
@@ -41,7 +41,7 @@ It is recommended to view the descriptions of each table to give you more contex
 
 # Understanding mapping outputs
 
-Running `[metadata_map](R/metadata_map.R)` will run this function in demo mode (as explained in the README) and generate four files in your project directory. 
+Running `[metadata_map](https://github.com/aim-rsf/mapmetadata/tree/main/R/metadata_map.R)` will run this function in demo mode (as explained in the README) and generate four files in your project directory. 
 
 By default, the demo mode processes the first 5 variables in a table. Running `metadata_map(demo_number = 20)` will process the first 20 variables in a table. 
 Here you can [view outputs generated from this longer demo run](https://github.com/aim-rsf/mapmetadata/tree/main/inst/outputs/), which include:
@@ -53,7 +53,7 @@ Here you can [view outputs generated from this longer demo run](https://github.c
 -   **MAPPING_360_NCCHD_CHILD_timestamp.csv**
     -   The mappings between variables in the CHILD table and the research domains.
 -   **L-MAPPING_360_NCCHD_CHILD_timestamp.csv**
-    -   The same mappings as the previous file, but saved in a longer format. See the argument `long_output = TRUE` in `[metadata_map](R/metadata_map.R)`. If `long_output = FALSE` was used, the long outputs can be generated by directly calling the `[map_convert](R/map_convert.R)` function. 
+    -   The same mappings as the previous file, but saved in a longer format. See the argument `long_output = TRUE` in `[metadata_map](https://github.com/aim-rsf/mapmetadata/tree/main/R/metadata_map.R)`. If `long_output = FALSE` was used, the long outputs can be generated by directly calling the `[map_convert](https://github.com/aim-rsf/mapmetadata/tree/main/R/map_convert.R)` function. 
 -   **MAPPING_LOG_360_NCCHD_CHILD_timestamp.csv**
     -   A log file that accompanies the MAPPING file, describing features of the session and the table processed.
 -   **MAPPING_PLOT_360_NCCHD_CHILD_timestamp.png**
@@ -61,11 +61,11 @@ Here you can [view outputs generated from this longer demo run](https://github.c
 
 # Compare mapping
 
-Running the function `[map_compare](R/map_compare.R)` will allow you to compare the mappings from two sessions, perhaps two different researchers. This function compares csv outputs from two sessions, finds their differences, and asks for a consensus, creating a new output file:
+Running the function `[map_compare](https://github.com/aim-rsf/mapmetadata/tree/main/R/map_compare.R)` will allow you to compare the mappings from two sessions, perhaps two different researchers. This function compares csv outputs from two sessions, finds their differences, and asks for a consensus, creating a new output file:
 
[…]
