### 2025-02-14 9543b8e
simplified the domain_list and re-ran example outputs
