### 2025-02-14 #674 @RayStick
Hi there, thanks again for the brilliant reviews. I have addressed them all (see this [issue](https://github.com/aim-rsf/mapmetadata/issues/184)) but I have not had a chance to do all my final sanity checks sorry for the delay. On Monday (with some fresh eyes) I will do some final checks then write a message on this thread clearly describing how I responded to each of your queries (following the guidance given). Thank you for your patience. The changes took much longer than I thought!
