### 2025-02-17 c90b04a
simplify the wa domain_list and look_up are used
