+++ R/metadata_map.R
@@ -15,12 +15,12 @@ select.list <- NULL
 #' variables, and categorisations for the same variable can be copied from one
 #' table to another. \cr \cr
 #' Example inputs are provided within the package data, for the user to run this
-#' function in a demo mode.
+#' function in a demo mode. Refer to the package website for more guidance.
 #' @param metadata_file This should be a csv download from HDRUK gateway
 #' (in the form of ID_Dataset_Metadata.csv). Run '?mapmetadata::metadata' to
 #' see how the metadata_file for the demo was created.
-#' @param domain_file This should be a csv file created by the user, with each
-#' domain on a separate line, no header. Run '?mapmetadata::domain_list' to
+#' @param domain_file This should be a csv file created by the user, with two
+#' columns (Domain_Code and Domain_Name). Run '?mapmetadata::domain_list' to
 #' see how the domain_file for the demo was created.
 #' @param look_up_file The lookup file makes auto-categorisations intended for
 #' variables that appear regularly in health datasets. It only works for 1:1
@@ -41,8 +41,8 @@ select.list <- NULL
 #' 'L-OUTPUT_' which gives each categorisation its own row. Default is TRUE.
 #' @param quiet Default is FALSE. Change to TRUE to quiet the cli_alert_info
 #' and cli_alert_success messages.
-#' @return The function will return two csv files: 'OUTPUT_' which contains the
-#' mappings and 'LOG_' which contains details about the dataset and session.
+#' @return A html plot summarising the dataset. Various csv and png outputs to
+#' summarise the user's mapping session for a specific table in the dataset.
 #' @examples
 #' # Demo run requires no function inputs but requires user interaction.
 #' # See package documentation to guide user inputs.
