### 2025-02-17 aim-rsf/mapmetadata#201 @RayStick
R2-4l (family tags)
**[l]** Function documentation sometimes includes references to other functions in the package, for example [`valid_comparison()`](https://github.com/aim-rsf/mapmetadata/blob/aea7f979af6b86c1654aba22722d9e43c557c505/R/compare.R#L3) refers to it being used in the `the map_compare.R function`. Cross links are mentioned in the [ROpenSci package guidelines](https://devguide.ropensci.org/pkg_building.html#roxygen-2-use) and if you use RStudio I recommend using my colleague, Matt Dray's, [snorkel](https://github.com/matt-dray/snorkel) package to add these in as you can highlight the text and use the RStudio add in.
