### 2025-02-17 41fcdeb
Update NEWS.md
