### 2025-02-17 aim-rsf/mapmetadata#184 @RayStick
@all-contributors add @ymansiaux for ideas and user testing
