### 2025-02-17 #674 @RayStick
My apologies, I was not too well today and after some user testing I found some bugs 😿 . Very close to finishing! Thanks for the patience.
