+++ vignettes/mapmetadata.Rmd
@@ -41,7 +41,7 @@ It is recommended to view the descriptions of each table to give you more contex
 
 # Understanding mapping outputs
 
-Running [metadata_map](https://aim-rsf.github.io/mapmetadata/reference/metadata_map.html) will run this function in demo mode (as explained in the README) and generate four files in your project directory. 
+Running [metadata_map](https://aim-rsf.github.io/mapmetadata/reference/metadata_map.html) will run this function in demo mode (as explained in the README) and generate six files in your project directory. 
 
 By default, the demo mode processes the first 5 variables in a table. Running `metadata_map(demo_number = 20)` will process the first 20 variables in a table. 
 Here you can [view outputs generated from this longer demo run](https://github.com/aim-rsf/mapmetadata/tree/main/inst/outputs/), which include:
@@ -166,6 +166,6 @@ If it finds other MAPPING files with the same ID_Name in the output directory, a
 Selection: 4
 ℹ Processing Table 4 of 13 (CHILD)
 
-ℹ Copying from previous session(s): MAPPING_360_NCCHD_CHILD_2025-02-14-18-14-01.csv
+ℹ Copying from previous session(s): MAPPING_360_NCCHD_BLOOD_TEST_2025-02-14-18-14-01.csv
 ```
-Instead of 'AUTO CATEGORISED' the note will say 'COPED FROM: CHILD' but you always have the option to manually override this copying during your review, or turn it off altogether with the option `table_copy = FASLE`.
+Instead of 'AUTO CATEGORISED' the note will say 'COPED FROM: BLOOD_TEST' but you always have the option to manually override this copying during your review, or turn it off altogether with the option `table_copy = FASLE`.
