### 2025-02-18 #674 @RayStick
# Response to reviewers

Please see the [NEWS file](https://aim-rsf.github.io/mapmetadata/news/index.html) and the new [4.0.0 release](https://github.com/aim-rsf/mapmetadata/releases/tag/v4.0.0) which incorporate all the changes. You can also see that [pkgcheck results](https://github.com/aim-rsf/mapmetadata/issues/188#issuecomment-2665091625) that were run on this 4.0.0 release showing everything passes

[This GitHub issue](https://github.com/aim-rsf/mapmetadata/issues/184) is where I itemized all your feedback and made changes in grouped PRs - that's where you can see all the detail. Below I will try to summarise for you in a more digestible way!

Please let me know if you have any further questions/suggestions/corrections.  Thank you again!

---

##  devchecks and lintr

This now has 0 errors, 0 warnings, and 1 note. The note is 'unable to verify current time' and I am not sure what to do about it!  I also ran `lintr::lint_package(path = ".")` and implemented the suggestions. This resolves both @ymansiaux and @Lextuga007 comments about devchecks(). This also resolves @Lextuga007 comments about 'Package build' and 'Formatting consistency'

##  Improvement to package code and tests

As suggested by @Lextuga007:
- Reduced the use of `cat()` and `print()` and allow suppression of some console messages ([PR](https://github.com/aim-rsf/mapmetadata/pull/193))
- Removed some requests from the user for console input (some you had challenges with) to simplify ([PR](https://github.com/aim-rsf/mapmetadata/pull/193))
- 'data element' = 'variable', sorry. I have removed all mentioned of 'data element' in the code & docs (a legacy term) and only say 'variable' now ([PR](https://github.com/aim-rsf/mapmetadata/issues/197))
- From reading your user testing comments about custom domain_list and lookup options, I realized I had over complicated this for a user. I have simplified the structure of these input files ([PR](https://github.com/aim-rsf/mapmetadata/pull/200/)) - this is all best explained by referencing the new README and package tutorial.

As suggested by @ymansiaux:
- Removed the loading of packages in `tests/testthat.R`  https://github.com/aim-rsf/mapmetadata/commit/986a56add28a2ae0a41c712f93bc1d43860d9e18
    - I struggled to replace my use of `{mockery}` with `testthat::local_mocked_bindings()` for [one of the unit tests](https://github.com/aim-rsf/mapmetadata/blob/main/tests/testthat/test-user_categorisation.R) - is that okay? Perhaps we can put as an enhancement in a GitHub Issue.
- Used `file.path()` for better compatibility across operating systems ([PR](https://github.com/aim-rsf/mapmetadata/issues/190))
- Improved the clarity of arguments used in the `map_convert()` function ([PR](https://github.com/aim-rsf/mapmetadata/pull/192/files))
- Implemented more extensive input validation ([PR](https://github.com/aim-rsf/mapmetadata/issues/197))

## Smaller documentation improvements: formatting, re-naming, clarifications

As suggested by @Lextuga007:
- Spelling corrections, improved accessibility of hyperlinks and explanation of technical language ([PR](https://github.com/aim-rsf/mapmetadata/issues/185)).
- Use the `@family` tag in function documentation, to group functions ([PR](https://github.com/aim-rsf/mapmetadata/issues/201)) and now link to the function references throughout the [tutorial](https://aim-rsf.github.io/mapmetadata/articles/mapmetadata.html).
- When referencing the package data I now use `?mapmetadata::metadata` and `data(metadata)` to make sure that users of this package (not in dev mode) can access these types of package files ([PR](https://github.com/aim-rsf/mapmetadata/issues/197))

As suggested by @ymansiaux:
- Prevented `.Rd` files from being generated in the `man/` folder by using the [devtag](https://github.com/moodymudskipper/devtag) as recommended by @maelle (https://github.com/aim-rsf/mapmetadata/commit/d180590bb09d0765784f68f9e04a6a33b3ca3b0f)
- Removed the `dontrun` tags from the Examples and used  `if(interactive())` where needed. Also used temporary directories for better clean up ([PR](https://github.com/aim-rsf/mapmetadata/pull/192))

## Larger documentation improvements

I have made lots of improvements to the [README](https://aim-rsf.github.io/mapmetadata/index.html) and the [longer tutorial](https://aim-rsf.github.io/mapmetadata/articles/mapmetadata.html), expanding, clarifying and re-structuring. This responds to a few comments from both reviewers. I made sure that the README only includes the *demo* run of the main function, and all code snippets that a user is prompted to run can be copied & pasted without modification. The demo run is much shorter now, it processes the first 5 variables by default (not 20 like before). In the [longer tutorial](https://aim-rsf.github.io/mapmetadata/articles/mapmetadata.html), I make it clearer to a wider range of package users how to access the example package files, and use these package files  to demo the custom inputs. Implementing more extensive input validation ([PR](https://github.com/aim-rsf/mapmetadata/issues/197)) and correcting some errors in the function docs also helps with user understanding I hope.

A few PRs but most notably - https://github.com/aim-rsf/mapmetadata/pull/196 and https://github.com/aim-rsf/mapmetadata/issues/199.
