### 2025-02-18 #674 @maelle
Thank you @RayStick!!

@ymansiaux @Lextuga007 could you please read @RayStick's response and answer using the [template](https://devguide.ropensci.org/approval2template.html)? Thank you!
