### 2025-02-18 #674 @maelle
@RayStick I'm intrigued by the mocking example. Where in the codebase is `readline()` introduced?
