### 2025-02-19 af4f309
test: use function factory instead of {mockery}
