### 2025-02-19 #674 @maelle
I made a suggestion in a PR https://github.com/aim-rsf/mapmetadata/pull/207 using function factories instead of mockery. Also curious to hear whether @ymansiaux or @Lextuga007 have any experience with this. It's also something you could ask in the rOpenSci slack workspace.
