### 2025-02-19 #674 @EPI-YMX
> I made a suggestion in a PR [aim-rsf/mapmetadata#207](https://github.com/aim-rsf/mapmetadata/pull/207) using function factories instead of mockery. Also curious to hear whether [@ymansiaux](https://github.com/ymansiaux) or [@Lextuga007](https://github.com/Lextuga007) have any experience with this. It's also something you could ask in the rOpenSci slack workspace.

Not really sorry. I remembered having used the trick `readline <- NULL` before, but your approach seems really interisting and more elegant !
