+++ tests/testthat/test-user_categorisation.R
@@ -1,18 +1,9 @@
 test_that("user_categorisation works with valid input", {
-  # create a mock object that returns user inputs in sequence
-  mock_factory <- function(...) {
-    i <- 0
 
-    function(...) {
-      things <- c("3", "This is a note", "n")
-      i <<- i + 1
-      things[i]
-    }
-  }
   # replace `readline` function within the `user_categorisation` function with
-  # the `mock_readline` mock object
+  # the `helper-mock_factory.R` mock object
   local_mocked_bindings(
-    readline = mock_factory()
+    readline = helper_mock("3", "This is a note", "n")
   )
 
   response <- user_categorisation(var = "Variable1",
@@ -22,11 +13,9 @@ test_that("user_categorisation works with valid input", {
 })
 
 test_that("user_categorisation handles invalid input and then valid input", {
-  # create a mock object that returns invalid input first, then valid input
-  mock_readline <- mockery::mock("6", "3", "This is a note", "n")
-  # replace `readline` function within the `user_categorisation` function with
-  # the `mock_readline` mock object
-  mockery::stub(user_categorisation, "readline", mock_readline)
+  local_mocked_bindings(
+    readline = helper_mock("6", "3", "This is a note", "n")
+  )
 
   response <- user_categorisation(var = "Variable1",
                                   desc = "Description1",
@@ -35,12 +24,9 @@ test_that("user_categorisation handles invalid input and then valid input", {
 })
 
 test_that("user_categorisation handles multiple valid inputs", {
-  # create a mock object that returns multiple valid inputs
-  mock_readline <- mockery::mock("3,4", "This is another note", "n")
-  # replace `readline` function within the `user_categorisation` function with
-  # the `mock_readline` mock object
-  mockery::stub(user_categorisation, "readline", mock_readline)
-
+  local_mocked_bindings(
+    readline = helper_mock("3,4", "This is another note", "n")
+  )
   response <- user_categorisation(var = "Variable1",
                                   desc = "Description1",
                                   type = "Type1", domain_code_max = 5)
@@ -49,12 +35,10 @@ test_that("user_categorisation handles multiple valid inputs", {
 })
 
 test_that("user_categorisation handles re-do input", {
-  # create a mock object that returns inputs including re-do
-  mock_readline <- mockery::mock("3", "This is a note", "y", "4",
-                                 "Another note", "n")
-  # replace `readline` function within the `user_categorisation` function with
-  # the `mock_readline` mock object
-  mockery::stub(user_categorisation, "readline", mock_readline)
+  local_mocked_bindings(
+    readline = helper_mock("3", "This is a note", "y", "4",
+                           "Another note", "n")
+  )
 
   response <- user_categorisation(var = "Variable1",
                                   desc = "Description1",
