### 2025-02-25 #674 @RayStick
Use of `mockery()` now removed: https://github.com/aim-rsf/mapmetadata/pull/207
Thanks @maelle =D
