### 2025-02-25 #674 @maelle
@Lextuga007 could you please read @RayStick's response and answer using the [template](https://devguide.ropensci.org/approval2template.html)? Thank you!
