### 2025-02-25 32c2485
re-arrange badges
