### 2025-03-03 #674 @Lextuga007
## Reviewer Response

Thank you so much for all the hard work you've put into the changes suggested!
So often small suggestions lead to major overhauls in code and you've handled this
well.
You also have documented these changes thoroughly!

I really enjoyed going through the README which has been substantially updated.
From a user experience it's a much quicker experience of getting started and I
really appreciated the details you've added to explain the data which can be
useful when someone is new to this as well as to R.

I've submitted a really small PR to edit just a couple of words in the messages.

A couple of minor enhancements/suggestions for the added content:

[`print`](https://github.com/aim-rsf/mapmetadata/blob/15a06c0f7f39c1f0f7b0fee7c1b60f11b3a5ecc0/vignettes/mapmetadata.Rmd#L119) is used to show the data in the Console - could this be just `look_up`?
Apologies too as I hadn't noticed this from my first review and this isn't new code.

```{r}
print(look_up, row.names = FALSE)

# and

look_up

# do the same printing but with less code
```

In the section "Using a custom lookup table input (advanced)" I wasn't sure if the appending suggestion of new rows to the bottom was expected to be done in the csv or in code.
The csv is easier but would require repeating steps which could be helpful to direct the user to.

> When creating a custom lookup table, you can use this lookup table as a base and just append new rows to the bottom.

<https://github.com/aim-rsf/mapmetadata/blob/15a06c0f7f39c1f0f7b0fee7c1b60f11b3a5ecc0/vignettes/mapmetadata.Rmd#L156>

As mentioned though, these are really small points and you've covered a huge amount in the review.
Thank you!

#### Final approval (post-review)

  Estimated hours spent reviewing: 2.5 hours
