### 2025-03-04 #674 @ropensci-review-bot
Approved! Thanks @RayStick for submitting and @ymansiaux, @Lextuga007 for your reviews! :grin:
