### 2025-03-04 #674 @RayStick
Wonderful! I have a bit of time today - I will take a look at @Lextuga007's comments and suggestions, as they seem like they would not take long. It would actually be great to transfer this to the rOpenSci GH Org as the current Org this repo is hosted on may not be supported in the months to come!
