### 2025-03-04 #674 @RayStick
Thanks @Lextuga007
All done:
- https://github.com/aim-rsf/mapmetadata/pull/212
- https://github.com/aim-rsf/mapmetadata/commit/d4b3bad88ae51b1727d840b845e714ee873ac541
- https://github.com/aim-rsf/mapmetadata/commit/5fee73d53c745eee759a527f719dfc20d7b33fe9
