### 2025-03-04 #674 @maelle
@RayStick it seems you haven't accepted the invitation to the ropensci GitHub organization yet. You should see it in your emails and I think in your GitHub notifications maybe.
